Create helper module for code management
