+++ assets/js/ccfw-admin-app.js
@@ -1,1 +1,1 @@
-(()=>{var e={615:(e,t,n)=>{"use strict";var r={add:'<path d="M24 10h-10v-10h-4v10h-10v4h10v10h4v-10h10z"/>',success:'<path d="M21.856 10.303c.086.554.144 1.118.144 1.697 0 6.075-4.925 11-11 11s-11-4.925-11-11 4.925-11 11-11c2.347 0 4.518.741 6.304 1.993l-1.422 1.457c-1.408-.913-3.082-1.45-4.882-1.45-4.962 0-9 4.038-9 9s4.038 9 9 9c4.894 0 8.879-3.928 8.99-8.795l1.866-1.902zm-.952-8.136l-9.404 9.639-3.843-3.614-3.095 3.098 6.938 6.71 12.5-12.737-3.096-3.096z"/>',check:'<path d="M9 21.035l-9-8.638 2.791-2.87 6.156 5.874 12.21-12.436 2.843 2.817z"/>',cross:'<path d="M23.954 21.03l-9.184-9.095 9.092-9.174-2.832-2.807-9.09 9.179-9.176-9.088-2.81 2.81 9.186 9.105-9.095 9.184 2.81 2.81 9.112-9.192 9.18 9.1z"/>',crossArrow:'<path d="M7 5h17v16h-17l-7-7.972 7-8.028zm7 6.586l-2.586-2.586-1.414 1.414 2.586 2.586-2.586 2.586 1.414 1.414 2.586-2.586 2.586 2.586 1.414-1.414-2.586-2.586 2.586-2.586-1.414-1.414-2.586 2.586z"/>',error:'<path d="M16.142 2l5.858 5.858v8.284l-5.858 5.858h-8.284l-5.858-5.858v-8.284l5.858-5.858h8.284zm.829-2h-9.942l-7.029 7.029v9.941l7.029 7.03h9.941l7.03-7.029v-9.942l-7.029-7.029zm-8.482 16.992l3.518-3.568 3.554 3.521 1.431-1.43-3.566-3.523 3.535-3.568-1.431-1.432-3.539 3.583-3.581-3.457-1.418 1.418 3.585 3.473-3.507 3.566 1.419 1.417z"/>',warning:'<path d="M12 0l-12 12 12 12 12-12-12-12zm-1 6h2v8h-2v-8zm1 12.25c-.69 0-1.25-.56-1.25-1.25s.56-1.25 1.25-1.25 1.25.56 1.25 1.25-.56 1.25-1.25 1.25z"/>',info:'<path d="M12 2c5.514 0 10 4.486 10 10s-4.486 10-10 10-10-4.486-10-10 4.486-10 10-10zm0-2c-6.627 0-12 5.373-12 12s5.373 12 12 12 12-5.373 12-12-5.373-12-12-12zm-2.033 16.01c.564-1.789 1.632-3.932 1.821-4.474.273-.787-.211-1.136-1.74.209l-.34-.64c1.744-1.897 5.335-2.326 4.113.613-.763 1.835-1.309 3.074-1.621 4.03-.455 1.393.694.828 1.819-.211.153.25.203.331.356.619-2.498 2.378-5.271 2.588-4.408-.146zm4.742-8.169c-.532.453-1.32.443-1.761-.022-.441-.465-.367-1.208.164-1.661.532-.453 1.32-.442 1.761.022.439.466.367 1.209-.164 1.661z"/>',bin:'<path d="M9 19c0 .552-.448 1-1 1s-1-.448-1-1v-10c0-.552.448-1 1-1s1 .448 1 1v10zm4 0c0 .552-.448 1-1 1s-1-.448-1-1v-10c0-.552.448-1 1-1s1 .448 1 1v10zm4 0c0 .552-.448 1-1 1s-1-.448-1-1v-10c0-.552.448-1 1-1s1 .448 1 1v10zm5-17v2h-20v-2h5.711c.9 0 1.631-1.099 1.631-2h5.315c0 .901.73 2 1.631 2h5.712zm-3 4v16h-14v-16h-2v18h18v-18h-2z"/>',save:'<path d="M13 3h2.996v5h-2.996v-5zm11 1v20h-24v-24h20l4 4zm-17 5h10v-7h-10v7zm15-4.171l-2.828-2.829h-.172v9h-14v-9h-3v20h20v-17.171zm-3 10.171h-14v1h14v-1zm0 2h-14v1h14v-1zm0 2h-14v1h14v-1z"/>',cookie:'<path d="M23.999 12.149c-.049 3.834-1.893 7.223-4.706 9.378-1.993 1.53-4.485 2.449-7.198 2.473-6.464.057-12.051-5.107-12.095-12 3.966 1.066 7.682-1.993 6-6 4.668.655 6.859-2.389 6.077-6 6.724.064 11.999 5.542 11.922 12.149zm-15.576-4.123c-.065 3.393-2.801 5.868-6.182 6.166 1.008 4.489 5.015 7.807 9.759 7.807 5.262 0 9.576-4.072 9.97-9.229.369-4.818-2.755-9.357-7.796-10.534-.277 2.908-2.381 5.357-5.751 5.79zm5.077 8.974c.828 0 1.5.672 1.5 1.5s-.672 1.5-1.5 1.5-1.5-.672-1.5-1.5.672-1.5 1.5-1.5zm-5.5-2.853c1.104 0 2 .896 2 2s-.896 2-2 2-2-.896-2-2 .896-2 2-2zm10-2.147c1.104 0 2 .896 2 2s-.896 2-2 2-2-.896-2-2 .896-2 2-2zm-5 0c.552 0 1 .448 1 1s-.448 1-1 1-1-.448-1-1 .448-1 1-1zm2.5-5c.828 0 1.5.672 1.5 1.5s-.672 1.5-1.5 1.5-1.5-.672-1.5-1.5.672-1.5 1.5-1.5zm-12.5 0c.552 0 1 .448 1 1s-.448 1-1 1-1-.448-1-1 .448-1 1-1zm-1.5-4c.828 0 1.5.672 1.5 1.5s-.672 1.5-1.5 1.5-1.5-.672-1.5-1.5.672-1.5 1.5-1.5zm6-2c.828 0 1.5.672 1.5 1.5s-.672 1.5-1.5 1.5-1.5-.672-1.5-1.5.672-1.5 1.5-1.5zm-3.5-1c.552 0 1 .448 1 1s-.448 1-1 1-1-.448-1-1 .448-1 1-1z"/>'},o={add:function(e,t){return i("add",e,t)},check:function(e,t){return i("check",e,t)},success:function(e,t){return i("success",e,t)},cross:function(e,t){return i("cross",e,t)},crossArrow:function(e,t){return i("crossArrow",e,t)},error:function(e,t){return i("error",e,t)},warning:function(e,t){return i("warning",e,t)},info:function(e,t){return i("info",e,t)},bin:function(e,t){return i("bin",e,t)},save:function(e,t){return i("save",e,t)},cookie:function(e,t){return i("cookie",e,t)}},i=function(e,t,n){switch(t=t||24,e){case"add":case"check":case"success":n=n||"#149414";break;case"cross":case"crossArrow":case"error":case"bin":n=n||"#CC0000";break;case"warning":n=n||"#FFA500";break;case"info":case"save":n=n||"#195e9f";break;case"cookie":n=n||"#984e0a";break;default:return'<b style="color:darkred">Error:</b> <b>Icon not found for <em>'+e+"</em></b>"}return'<svg xmlns="http://www.w3.org/2000/svg" class="ccfw-icon ccfw-icon__'+e+'" width="'+t+'" height="'+t+'" fill="'+n+'" viewBox="0 0 24 24">'+r[e]+" </svg>"};var a=function(e){return function(e){jQuery.ajax({type:"POST",url:ajaxurl,data:{action:"ccfw_cookie_store",payload:e},beforeSend:function(){jQuery(".ccfw-ajax").addClass("active").fadeIn("fast")}}).done((function(){jQuery(".ccfw-ajax").fadeOut("fast",(function(){jQuery(this).removeClass("active")})),C.sectionsStored=!0,jQuery("#submit").click()})).fail((function(e){console.log(e)}))}(e)},c=function(){jQuery.ajax({url:ajaxurl,data:{action:"ccfw_cookie_get"}}).done((function(e){C.sections=JSON.parse(e)||{},Object.keys(C.sections).length>0&&(C.sectionsLoaded=!0),C.debug.output()}))},s=function(e,t){return jQuery("<"+e+"/>",t||null)},u=function(e){return s("span",{class:"ccfw-group__heading"}).html("Group name: <span>"+(e||"")+"</span>")};function p(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=null==e?null:"undefined"!=typeof Symbol&&e[Symbol.iterator]||e["@@iterator"];if(null==n)return;var r,o,i=[],a=!0,c=!1;try{for(n=n.call(e);!(a=(r=n.next()).done)&&(i.push(r.value),!t||i.length!==t);a=!0);}catch(e){c=!0,o=e}finally{try{a||null==n.return||n.return()}finally{if(c)throw o}}return i}(e,t)||function(e,t){if(!e)return;if("string"==typeof e)return f(e,t);var n=Object.prototype.toString.call(e).slice(8,-1);"Object"===n&&e.constructor&&(n=e.constructor.name);if("Map"===n||"Set"===n)return Array.from(e);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return f(e,t)}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function f(e,t){(null==t||t>e.length)&&(t=e.length);for(var n=0,r=new Array(t);n<t;n++)r[n]=e[n];return r}var d={init:function(){L.cookies.get();var e=setInterval((function(){C.sectionsLoaded&&(clearInterval(e),d.load())}),100)},load:function(){C.building=!0;for(var e=0,t=Object.entries(C.sections);e<t.length;e++){var n=p(t[e],2),r=n[0],o=n[1];if(C.create.section(r),o)for(var i=0,a=Object.entries(o);i<a.length;i++){var c=p(a[i],2),s=(c[0],c[1]);if(C.create.group({section:r,group:s.name,description:s.description,id:s.allowlistID}),s.cookies)for(var u=0,f=Object.entries(s.cookies);u<f.length;u++){var d=p(f[u],2),l=d[0],v=d[1];C.create.cookie(l.replace("row_",""),{section:r,group:s.name,row:v})}}}C.building=!1}};function l(){var e=jQuery(this).is(":checked"),t=jQuery("#"+C.debug.container);e?(C.debug.active=!0,0===t.length&&C.debug.make(),t.fadeIn()):(C.debug.active=!1,t.fadeOut())}var v=function(){var e=jQuery("."+C.debug.import.textarea).val();if(C.manage.isJson(e)){var t=JSON.parse(e);t&&(C.sections=t,d.load(),C.debug.output())}};function m(){var e=jQuery(this),t=e.text(),n=jQuery("<input>");jQuery("body").append(n),n.val(t).select(),document.execCommand("copy"),n.remove(),e.addClass("copied"),setTimeout((function(){return e.removeClass("copied")}),500)}var h={save:function(e,t){return function(e,t){var n=jQuery(".ccfw-"+e+"-group-save");n.off("click",t),n.on("click",t)}(e,t)},remove:function(e){return function(e){var t=jQuery(".ccfw-group-remove");t.off("click",e),t.on("click",e)}(e)},allowlist:{save:function(e){return function(e){jQuery(".ccfw-cookie-row__gtm-allowlist-id").on("keyup",e)}(e)}},description:{save:function(e){return function(e){jQuery(".ccfw-group__description input").on("keyup",e)}(e)}}},g={add:function(e){return function(e){var t=jQuery(".ccfw-cookie-row-add");t.off("click",e),t.on("click",e)}(e)},save:function(e){return function(e){var t=jQuery(".ccfw-cookie-row input");t.off("keyup",e),t.on("keyup",e)}(e)},remove:function(e){return function(e){var t=jQuery(".ccfw-cookie-row-remove");t.off("click",e),t.on("click",e)}(e)}},b={submit:function(e){return jQuery("#ccfw-form").on("submit",e)},input:{focusBlur:function(e){return jQuery("#"+C.appContainer).on("focus blur","input",e)}}},y={toggle:function(){jQuery("#"+C.debug.checkbox).on("change",l)},import:function(){return(e=jQuery("."+C.debug.import.textarea)).on("keypress",C.manage.readonly),void e.on("keyup",v);var e},copy:function(){jQuery("#"+C.debug.preContainer).find("pre").on("click",m)}},w=function(e){var t=e||"-- Create an item --";return s("option",{value:(e=e||"").toLowerCase()}).text(t)},x=function(e,t){return s("input",{type:"text",name:e,placeholder:t||"",class:"ccfw-cookie-row__"+e})},k=function(e,t,n,r){var o=s("button",{type:"button",class:e});return n&&o.append(n),r&&o.attr("title",r),t&&o.append(t),o};function O(e){return C.manage.allowlistIdsValid()?C.sectionsChanged&&(removeEventListener("beforeunload",C.manage.beforeUnload,{capture:!0}),jQuery("html, body").animate({scrollTop:30},"fast"),!C.sectionsStored)?(e.preventDefault(),L.form.post(C.sections),!1):void 0:(e.preventDefault(),!1)}function j(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=null==e?null:"undefined"!=typeof Symbol&&e[Symbol.iterator]||e["@@iterator"];if(null==n)return;var r,o,i=[],a=!0,c=!1;try{for(n=n.call(e);!(a=(r=n.next()).done)&&(i.push(r.value),!t||i.length!==t);a=!0);}catch(e){c=!0,o=e}finally{try{a||null==n.return||n.return()}finally{if(c)throw o}}return i}(e,t)||function(e,t){if(!e)return;if("string"==typeof e)return E(e,t);var n=Object.prototype.toString.call(e).slice(8,-1);"Object"===n&&e.constructor&&(n=e.constructor.name);if("Map"===n||"Set"===n)return Array.from(e);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return E(e,t)}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function E(e,t){(null==t||t>e.length)&&(t=e.length);for(var n=0,r=new Array(t);n<t;n++)r[n]=e[n];return r}var C={appContainer:"ccfw-cookie-management",sectionId:"ccfw-section-select-container",sections:{},sectionsStored:!1,sectionsLoaded:!1,sectionsChanged:!1,building:!1,cookieRow:{name:"",description:"",expiry:""},debug:{container:"debugging",preContainer:"debug-settings",import:{container:"ccfw-import-container",textarea:"ccfw-import-cookie-object"},active:!1,checkbox:"cookie-management-debug",init:function(){jQuery.migrateMute=!0,jQuery("#"+C.debug.checkbox).is(":checked")&&(C.debug.active=!0,C.debug.make())},make:function(){var e,t,n=s("div",{id:C.debug.container}),r=s("div",{class:C.debug.container+"-icons"});r.append((t="<h3>Available icons</h3>",e=(e=8)||8,Object.entries(o).forEach((function(n,r){t+=n[1]()+((r+1)%e==0?"<br><br>":" &nbsp; ")})),t)),n.append(r),n.append('<div id ="'+C.debug.preContainer+'"/>'),jQuery("#"+C.appContainer).before(n),C.debug.output()},output:function(){if(C.debug.active){var e=JSON.stringify(C.sections,void 0,4);if(e.length<5){var t=s("div",{id:C.debug.import.container});t.append("<h3>Import Cookies</h3>"),t.append(s("textarea",{class:C.debug.import.textarea})),jQuery("#"+C.debug.preContainer).html(t),y.import()}else{document.getElementById(C.debug.preContainer).innerHTML="<pre>"+e+"</pre>",y.copy()}}}},alert:function(e){function t(t){return e.apply(this,arguments)}return t.toString=function(){return e.toString()},t}((function(e){alert(e)})),confirm:function(e){return window.confirm(e)},create:{app:function(){var e=jQuery("#"+C.appContainer);e.html(C.manage.sectionSelect().append(C.manage.appInfoButton())),e.prepend('<div class="ccfw-ajax"> <div class="loader"> </div> </div>')},section:function(e){var t=C.building?"#"+C.sectionId+"-select":this,n=jQuery(t),r=C.building?e:n.val(),i=s("div",{id:"ccfw-section__"+r,class:"ccfw-section"}).data("id",r),a=s("button",{type:"button",id:"ccfw-"+r+"-section-remove",class:"ccfw-section-remove",title:"Click to remove the "+r+" section and all related cookies"});L.section.add(r),a.append(o.bin("20")),i.append(a),i.append(s("h2").text(S(r)+" Cookies"));var c=s("div",{class:"ccfw-group"}),p=s("div",{class:"ccfw-group-define"}),f=s("button",{type:"button",class:"ccfw-"+r+"-group-save",title:"Create a group"});f.append(o.add(22)),c.append(u()),p.append(x("ccfw-group__input")),p.append(f),c.append(p),i.append(c),jQuery("#"+C.appContainer).append(i),C.building||c.find('input[name="ccfw-group__input"]').focus(),h.save(r,C.create.group),jQuery("#ccfw-"+r+"-section-remove").on("click",C.manage.deleteSection),n.find("option[value="+r+"]").remove(),C.manage.hideShowSectionSelect()},group:function(e){var t=C.building?".ccfw-"+e.section+"-group-save":this,n=jQuery(t).closest("div.ccfw-section"),r=jQuery(t).closest("div.ccfw-group"),i=C.building?e.group:r.find("input").val(),a=D(i),c=n.data("id"),p="",f="marketing"===c?x("gtm-allowlist-id","GTM Allowlist ID"):null,d=x("ccfw-group-description","Enter "+i+" group blurb");if(i)if(C.building||!L.group.exists(c,a)){L.group.add(c,i),r.data("id",a),r.addClass("ccfw-group__"+a),r.html(u(i+" "+o.success(18))),r.append(k("ccfw-group-remove",o.bin(16,"grey"))),f&&r.append(f),(p=s("div",{class:"ccfw-group__description"})).append(d),p.append(k("ccfw-cookie-row-add ccfw-cookie-row-add-"+a,o.add(20),"Add cookie")),r.append(p);var l=s("div",{class:"ccfw-cookie-container"});if(!C.building){var v=C.create.cookieRow(0);l.append(v),L.cookies.add(c,i,"row_0"),v.find('input[name="name"]').focus()}r.append(l),n.append(C.manage.group(c)),h.save(c,C.create.group),h.remove(C.manage.removeGroup),h.allowlist.save(C.save.allowListId),h.description.save(C.save.description),g.add(C.create.cookie),g.remove(C.manage.removeRow),g.save(C.save.cookieData),C.building&&(f&&f.val(e.id),d.val(e.description))}else C.alert('The group "'+i+'" already exists. Please chose another name.');else console.warn("Group name was empty.")},cookie:function(e,t){var n=C.building?"#ccfw-section__"+t.section+" .ccfw-cookie-row-add-"+D(t.group):this,r=jQuery(n).parent().siblings(".ccfw-cookie-container"),o=C.building?e:r.find(".ccfw-cookie-row").length,i=C.manage.locationData(jQuery(n)),a=!1;if(!C.building)for(;!1===a;){if(!L.cookies.row.exists(i.section,i.group,"row_"+o)){a=!0;break}o++}var c=C.create.cookieRow(o);r.append(c),C.building?(c.find('input[name="name"]').val(t.row.name),c.find('input[name="description"]').val(t.row.description),c.find('input[name="expiry"]').val(t.row.expiry)):c.find('input[name="name"]').focus(),L.cookies.add(i.section,i.group,"row_"+o),g.save(C.save.cookieData),g.remove(C.manage.removeRow)},cookieRow:function(e){var t=s("div",{class:"ccfw-cookie-row"}).data("id","row_"+e);return t.append(x("name","Name")),t.append(x("description","Description")),t.append(x("expiry","Expiry or URL")),t.append(k("ccfw-cookie-row-remove",o.crossArrow(30))),t}},save:{allowListId:function(){var e=jQuery(this),t=C.manage.locationData(e);e.removeClass("ccfw-form-input-invalid"),L.group.allowlistId.save(t,e.val())},description:function(){var e=jQuery(this),t=C.manage.locationData(e);L.group.description.save(t,e.val())},cookieData:function(){var e=jQuery(this),t=C.manage.locationData(e),n=e.parent().data("id"),r=e.prop("name");L.cookies.update(t.section,t.group,n,r,e.val())}},manage:{hideShowSectionSelect:function(){jQuery("#"+C.sectionId+" option").length>1?jQuery("#"+C.sectionId).show():jQuery("#"+C.sectionId).hide()},deleteSection:function(){var e=jQuery(this).parent("div").data("id");C.confirm("\nAre you sure you want to remove "+e+" cookies?")&&(jQuery("#ccfw-section__"+e).remove(),L.section.remove(e),jQuery("#"+C.sectionId+" select").append(w(S(e))),C.manage.hideShowSectionSelect())},inputFocus:function(e){var t=jQuery(this).closest(".ccfw-group");"focusin"===e.type?t.removeClass("ccfw-group__active-opacity"):t.addClass("ccfw-group__active-opacity")},locationData:function(e){return{section:e.closest("div.ccfw-section").data("id"),group:e.closest("div.ccfw-group").data("id")}},appInfoButton:function(){return k("app-info",o.info(),null,"Choose an option from the dropdown to create a cookie section. The section will allow you to define groups of cookies.")},removeGroup:function(){var e=C.manage.locationData(jQuery(this));C.confirm("\nAre you sure you want to remove the "+e.group+" group from "+e.section+" cookies?")&&(jQuery(".ccfw-group__"+e.group).remove(),L.group.remove(e.section,e.group))},removeRow:function(){var e=C.manage.locationData(jQuery(this)),t=jQuery(this).parent(),n=t.data("id");L.cookies.remove(e.section,e.group,n),t.remove()},group:function(e){var t=s("div",{class:"ccfw-group"}),n=s("div",{class:"ccfw-group-define"}),r=s("button",{type:"button",class:"ccfw-"+e+"-group-save",title:"Create a group"});return r.append(o.add(22)),t.append(u()),n.append(x("ccfw-group__input")),n.append(r),t.append(n),t},sectionSelect:function(e){var t=["Third-party","Essential","Marketing"];return e&&e.forEach((function(e){t.push(e)})),function(e,t,n){var r=s("select",{name:e,id:e+"-select"}),o=s("div",{id:e});return r.append(w("Create a section")),t.forEach((function(e){r.append(w(e))})),r.on("change",n),o.html(r),o}(C.sectionId,t.reverse(),C.create.section)},beforeUnload:function(e){return e.preventDefault(),e.returnValue="Changes have been made to cookie content. Are you sure you want to leave?"},allowlistIdsValid:function(){var e=!0;console.log("Checking IDs....");for(var t=0,n=Object.entries(C.sections);t<n.length;t++){var r=j(n[t],2),o=r[0],i=r[1];if("marketing"===o)for(var a=0,c=Object.entries(i);a<c.length;a++){var s=j(c[a],2),u=s[0],p=s[1];T(o,u)||(jQuery("#ccfw-section__"+o+" .ccfw-group__"+u+" .ccfw-cookie-row__gtm-allowlist-id").addClass("ccfw-form-input-invalid"),C.alert("The "+p.name+" group has an invalid GTM Allowlist ID."),e=!1)}}return e},readonly:function(e){var t=45,n=46,r=97,o=99,i=118,a=120,c=121,s=122,u=e.which;if(e.ctrlKey&&u==c||e.ctrlKey&&u==s||e.ctrlKey&&u==a||e.ctrlKey&&u==i||e.ctrlKey&&u==o||e.shiftKey&&e.keyCode==t||e.shiftKey&&e.keyCode==n||e.ctrlKey&&e.keyCode==t||e.ctrlKey&&u==r)return 0;var p=e||window.event,f=p.keyCode||p.which;f=String.fromCharCode(f);/[]|\./.test(f)||(p.returnValue=!1,p.preventDefault())},isJson:function(e){try{return[null,JSON.parse(e)]}catch(e){}}}},A=function(){C.sectionsChanged=!0,addEventListener("beforeunload",C.manage.beforeUnload,{capture:!0})},T=function(e,t){var n=C.sections[e][t];return!!n.hasOwnProperty("allowlistID")&&0!==n.allowlistID.length},S=function(e){return"string"!=typeof e?"":e.charAt(0).toUpperCase()+e.slice(1)},D=function(e){return e.toLowerCase().replace(/ /g,"-").replace(/[-]+/g,"-").replace(/[^\w-]+/g,"")},L={section:{add:function(e){if(C.building)return!1;C.sections[e]={},A(),C.debug.output()},remove:function(e){if(C.building)return!1;delete C.sections[e],A(),C.debug.output()}},group:{add:function(e,t){if(C.building)return!1;A();var n=D(t);C.sections[e][n]={name:t},C.sections[e][n].cookies={row_0:{name:"",description:"",expiry:""}},C.debug.output()},remove:function(e,t){if(C.building)return!1;A(),delete C.sections[e][t],C.debug.output()},exists:function(e,t){return C.sections[e].hasOwnProperty(t)},allowlistId:{save:function(e,t){if(C.building)return!1;A(),C.sections[e.section][D(e.group)].allowlistID=t.trim(),C.debug.output()}},description:{save:function(e,t){if(C.building)return!1;A(),C.sections[e.section][D(e.group)].description=t.trim(),C.debug.output()}}},cookies:{add:function(e,t,n){if(C.building)return!1;A(),C.sections[e][D(t)].cookies[n]={name:"",description:"",expiry:""},C.debug.output()},update:function(e,t,n,r,o){if(C.building)return!1;A(),C.sections[e][D(t)].cookies[n][r]=o.trim(),C.debug.output()},remove:function(e,t,n){if(C.building)return!1;A(),delete C.sections[e][D(t)].cookies[n],C.debug.output()},row:{exists:function(e,t,n){return t=D(t),C.sections[e][t].cookies||(C.sections[e][t].cookies={}),C.sections[e][t].cookies.hasOwnProperty(n)}},get:function(){return c()}},form:{post:function(e){return a(e)}}},_=n(861),I=n(379),M=n.n(I),Q=n(535),z={insert:"head",singleton:!1};M()(Q.Z,z);Q.Z.locals;jQuery((function(){d.init(),C.debug.init(),b.input.focusBlur(C.manage.inputFocus),y.toggle(),b.submit(O),C.create.app(),jQuery("h2:contains(Cookie Management Settings)").css({marginBottom:"1.68em"}).prepend(o.cookie(28)),(0,_.ZP)(".app-info",{aria:{content:"describedby"},content:function(e){return e.getAttribute("title")},offset:[-18,10]})}))},502:()=>{function e(t){return(e="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(t)}Object.keys||(Object.keys=function(){"use strict";var t=Object.prototype.hasOwnProperty,n=!{toString:null}.propertyIsEnumerable("toString"),r=["toString","toLocaleString","valueOf","hasOwnProperty","isPrototypeOf","propertyIsEnumerable","constructor"],o=r.length;return function(i){if("function"!=typeof i&&("object"!==e(i)||null===i))throw new TypeError("Object.keys called on non-object");var a,c,s=[];for(a in i)t.call(i,a)&&s.push(a);if(n)for(c=0;c<o;c++)t.call(i,r[c])&&s.push(r[c]);return s}}())},108:()=>{jQuery((function(e){function t(e,t){var n=new URLSearchParams(window.location.search);return!t&&n.has(e)?n.get(e):!!e&&(n.set(e,t),window.history&&window.history.replaceState({},"","".concat(location.pathname,"?").concat(n)),window.location.pathname+window.location.search)}function n(t){return t.preventDefault(),r(e(this).attr("href")),!1}function r(n){var r,o,i=e(".nav-tab-wrapper a");return(n=n?e(".nav-tab-wrapper a[href='"+n+"']"):i.eq(0)).attr("href")||(n=i.eq(0)),r=n.attr("href").split("#")[1],n.parent().find("a").removeClass("nav-tab-active"),n.addClass("nav-tab-active"),e(".ccfw-component-settings-section").hide(),e("div#"+r).fadeIn(),o=t("ccfw-tab",r),e('input[name="_wp_http_referer"]').val(o),!1}if(e(".settings_page_cookie-compliance-for-wordpress-settings").length>0){e(".nav-tab-wrapper").on("click","a",n),e(".ccfw-inline-tab-link").on("click",n);var o=t("ccfw-tab");r(o?"#"+o:null)}})),Object.entries||(Object.entries=function(e){for(var t=Object.keys(e),n=t.length,r=new Array(n);n--;)r[n]=[t[n],e[t[n]]];return r})},535:(e,t,n)=>{"use strict";n.d(t,{Z:()=>i});var r=n(645),o=n.n(r)()((function(e){return e[1]}));o.push([e.id,'.tippy-box[data-animation=fade][data-state=hidden]{opacity:0}[data-tippy-root]{max-width:calc(100vw - 10px)}.tippy-box{background-color:#333;border-radius:4px;color:#fff;font-size:14px;line-height:1.4;outline:0;position:relative;transition-property:transform,visibility,opacity}.tippy-box[data-placement^=top]>.tippy-arrow{bottom:0}.tippy-box[data-placement^=top]>.tippy-arrow:before{border-top-color:initial;border-width:8px 8px 0;bottom:-7px;left:0;transform-origin:center top}.tippy-box[data-placement^=bottom]>.tippy-arrow{top:0}.tippy-box[data-placement^=bottom]>.tippy-arrow:before{border-bottom-color:initial;border-width:0 8px 8px;left:0;top:-7px;transform-origin:center bottom}.tippy-box[data-placement^=left]>.tippy-arrow{right:0}.tippy-box[data-placement^=left]>.tippy-arrow:before{border-left-color:initial;border-width:8px 0 8px 8px;right:-7px;transform-origin:center left}.tippy-box[data-placement^=right]>.tippy-arrow{left:0}.tippy-box[data-placement^=right]>.tippy-arrow:before{border-right-color:initial;border-width:8px 8px 8px 0;left:-7px;transform-origin:center right}.tippy-box[data-inertia][data-state=visible]{transition-timing-function:cubic-bezier(.54,1.5,.38,1.11)}.tippy-arrow{color:#333;height:16px;width:16px}.tippy-arrow:before{border-color:transparent;border-style:solid;content:"";position:absolute}.tippy-content{padding:5px 9px;position:relative;z-index:1}',""]);const i=o},645:e=>{"use strict";e.exports=function(e){var t=[];return t.toString=function(){return this.map((function(t){var n=e(t);return t[2]?"@media ".concat(t[2]," {").concat(n,"}"):n})).join("")},t.i=function(e,n,r){"string"==typeof e&&(e=[[null,e,""]]);var o={};if(r)for(var i=0;i<this.length;i++){var a=this[i][0];null!=a&&(o[a]=!0)}for(var c=0;c<e.length;c++){var s=[].concat(e[c]);r&&o[s[0]]||(n&&(s[2]?s[2]="".concat(n," and ").concat(s[2]):s[2]=n),t.push(s))}},t}},379:(e,t,n)=>{"use strict";var r,o=function(){return void 0===r&&(r=Boolean(window&&document&&document.all&&!window.atob)),r},i=function(){var e={};return function(t){if(void 0===e[t]){var n=document.querySelector(t);if(window.HTMLIFrameElement&&n instanceof window.HTMLIFrameElement)try{n=n.contentDocument.head}catch(e){n=null}e[t]=n}return e[t]}}(),a=[];function c(e){for(var t=-1,n=0;n<a.length;n++)if(a[n].identifier===e){t=n;break}return t}function s(e,t){for(var n={},r=[],o=0;o<e.length;o++){var i=e[o],s=t.base?i[0]+t.base:i[0],u=n[s]||0,p="".concat(s," ").concat(u);n[s]=u+1;var f=c(p),d={css:i[1],media:i[2],sourceMap:i[3]};-1!==f?(a[f].references++,a[f].updater(d)):a.push({identifier:p,updater:h(d,t),references:1}),r.push(p)}return r}function u(e){var t=document.createElement("style"),r=e.attributes||{};if(void 0===r.nonce){var o=n.nc;o&&(r.nonce=o)}if(Object.keys(r).forEach((function(e){t.setAttribute(e,r[e])})),"function"==typeof e.insert)e.insert(t);else{var a=i(e.insert||"head");if(!a)throw new Error("Couldn't find a style target. This probably means that the value for the 'insert' parameter is invalid.");a.appendChild(t)}return t}var p,f=(p=[],function(e,t){return p[e]=t,p.filter(Boolean).join("\n")});function d(e,t,n,r){var o=n?"":r.media?"@media ".concat(r.media," {").concat(r.css,"}"):r.css;if(e.styleSheet)e.styleSheet.cssText=f(t,o);else{var i=document.createTextNode(o),a=e.childNodes;a[t]&&e.removeChild(a[t]),a.length?e.insertBefore(i,a[t]):e.appendChild(i)}}function l(e,t,n){var r=n.css,o=n.media,i=n.sourceMap;if(o?e.setAttribute("media",o):e.removeAttribute("media"),i&&"undefined"!=typeof btoa&&(r+="\n/*# sourceMappingURL=data:application/json;base64,".concat(btoa(unescape(encodeURIComponent(JSON.stringify(i))))," */")),e.styleSheet)e.styleSheet.cssText=r;else{for(;e.firstChild;)e.removeChild(e.firstChild);e.appendChild(document.createTextNode(r))}}var v=null,m=0;function h(e,t){var n,r,o;if(t.singleton){var i=m++;n=v||(v=u(t)),r=d.bind(null,n,i,!1),o=d.bind(null,n,i,!0)}else n=u(t),r=l.bind(null,n,t),o=function(){!function(e){if(null===e.parentNode)return!1;e.parentNode.removeChild(e)}(n)};return r(e),function(t){if(t){if(t.css===e.css&&t.media===e.media&&t.sourceMap===e.sourceMap)return;r(e=t)}else o()}}e.exports=function(e,t){(t=t||{}).singleton||"boolean"==typeof t.singleton||(t.singleton=o());var n=s(e=e||[],t);return function(e){if(e=e||[],"[object Array]"===Object.prototype.toString.call(e)){for(var r=0;r<n.length;r++){var o=c(n[r]);a[o].references--}for(var i=s(e,t),u=0;u<n.length;u++){var p=c(n[u]);0===a[p].references&&(a[p].updater(),a.splice(p,1))}n=i}}}},861:(e,t,n)=>{"use strict";function r(e){var t=e.getBoundingClientRect();return{width:t.width,height:t.height,top:t.top,right:t.right,bottom:t.bottom,left:t.left,x:t.left,y:t.top}}function o(e){if(null==e)return window;if("[object Window]"!==e.toString()){var t=e.ownerDocument;return t&&t.defaultView||window}return e}function i(e){var t=o(e);return{scrollLeft:t.pageXOffset,scrollTop:t.pageYOffset}}function a(e){return e instanceof o(e).Element||e instanceof Element}function c(e){return e instanceof o(e).HTMLElement||e instanceof HTMLElement}function s(e){return"undefined"!=typeof ShadowRoot&&(e instanceof o(e).ShadowRoot||e instanceof ShadowRoot)}function u(e){return e?(e.nodeName||"").toLowerCase():null}function p(e){return((a(e)?e.ownerDocument:e.document)||window.document).documentElement}function f(e){return r(p(e)).left+i(e).scrollLeft}function d(e){return o(e).getComputedStyle(e)}function l(e){var t=d(e),n=t.overflow,r=t.overflowX,o=t.overflowY;return/auto|scroll|overlay|hidden/.test(n+o+r)}function v(e,t,n){void 0===n&&(n=!1);var a,s,d=p(t),v=r(e),m=c(t),h={scrollLeft:0,scrollTop:0},g={x:0,y:0};return(m||!m&&!n)&&(("body"!==u(t)||l(d))&&(h=(a=t)!==o(a)&&c(a)?{scrollLeft:(s=a).scrollLeft,scrollTop:s.scrollTop}:i(a)),c(t)?((g=r(t)).x+=t.clientLeft,g.y+=t.clientTop):d&&(g.x=f(d))),{x:v.left+h.scrollLeft-g.x,y:v.top+h.scrollTop-g.y,width:v.width,height:v.height}}function m(e){var t=r(e),n=e.offsetWidth,o=e.offsetHeight;return Math.abs(t.width-n)<=1&&(n=t.width),Math.abs(t.height-o)<=1&&(o=t.height),{x:e.offsetLeft,y:e.offsetTop,width:n,height:o}}function h(e){return"html"===u(e)?e:e.assignedSlot||e.parentNode||(s(e)?e.host:null)||p(e)}function g(e){return["html","body","#document"].indexOf(u(e))>=0?e.ownerDocument.body:c(e)&&l(e)?e:g(h(e))}function b(e,t){var n;void 0===t&&(t=[]);var r=g(e),i=r===(null==(n=e.ownerDocument)?void 0:n.body),a=o(r),c=i?[a].concat(a.visualViewport||[],l(r)?r:[]):r,s=t.concat(c);return i?s:s.concat(b(h(c)))}function y(e){return["table","td","th"].indexOf(u(e))>=0}function w(e){return c(e)&&"fixed"!==d(e).position?e.offsetParent:null}function x(e){for(var t=o(e),n=w(e);n&&y(n)&&"static"===d(n).position;)n=w(n);return n&&("html"===u(n)||"body"===u(n)&&"static"===d(n).position)?t:n||function(e){var t=-1!==navigator.userAgent.toLowerCase().indexOf("firefox");if(-1!==navigator.userAgent.indexOf("Trident")&&c(e)&&"fixed"===d(e).position)return null;for(var n=h(e);c(n)&&["html","body"].indexOf(u(n))<0;){var r=d(n);if("none"!==r.transform||"none"!==r.perspective||"paint"===r.contain||-1!==["transform","perspective"].indexOf(r.willChange)||t&&"filter"===r.willChange||t&&r.filter&&"none"!==r.filter)return n;n=n.parentNode}return null}(e)||t}n.d(t,{ZP:()=>rt});var k="top",O="bottom",j="right",E="left",C="auto",A=[k,O,j,E],T="start",S="end",D="viewport",L="popper",_=A.reduce((function(e,t){return e.concat([t+"-"+T,t+"-"+S])}),[]),I=[].concat(A,[C]).reduce((function(e,t){return e.concat([t,t+"-"+T,t+"-"+S])}),[]),M=["beforeRead","read","afterRead","beforeMain","main","afterMain","beforeWrite","write","afterWrite"];function Q(e){var t=new Map,n=new Set,r=[];function o(e){n.add(e.name),[].concat(e.requires||[],e.requiresIfExists||[]).forEach((function(e){if(!n.has(e)){var r=t.get(e);r&&o(r)}})),r.push(e)}return e.forEach((function(e){t.set(e.name,e)})),e.forEach((function(e){n.has(e.name)||o(e)})),r}var z={placement:"bottom",modifiers:[],strategy:"absolute"};function P(){for(var e=arguments.length,t=new Array(e),n=0;n<e;n++)t[n]=arguments[n];return!t.some((function(e){return!(e&&"function"==typeof e.getBoundingClientRect)}))}function B(e){void 0===e&&(e={});var t=e,n=t.defaultModifiers,r=void 0===n?[]:n,o=t.defaultOptions,i=void 0===o?z:o;return function(e,t,n){void 0===n&&(n=i);var o,c,s={placement:"bottom",orderedModifiers:[],options:Object.assign({},z,i),modifiersData:{},elements:{reference:e,popper:t},attributes:{},styles:{}},u=[],p=!1,f={state:s,setOptions:function(n){d(),s.options=Object.assign({},i,s.options,n),s.scrollParents={reference:a(e)?b(e):e.contextElement?b(e.contextElement):[],popper:b(t)};var o=function(e){var t=Q(e);return M.reduce((function(e,n){return e.concat(t.filter((function(e){return e.phase===n})))}),[])}(function(e){var t=e.reduce((function(e,t){var n=e[t.name];return e[t.name]=n?Object.assign({},n,t,{options:Object.assign({},n.options,t.options),data:Object.assign({},n.data,t.data)}):t,e}),{});return Object.keys(t).map((function(e){return t[e]}))}([].concat(r,s.options.modifiers)));return s.orderedModifiers=o.filter((function(e){return e.enabled})),s.orderedModifiers.forEach((function(e){var t=e.name,n=e.options,r=void 0===n?{}:n,o=e.effect;if("function"==typeof o){var i=o({state:s,name:t,instance:f,options:r}),a=function(){};u.push(i||a)}})),f.update()},forceUpdate:function(){if(!p){var e=s.elements,t=e.reference,n=e.popper;if(P(t,n)){s.rects={reference:v(t,x(n),"fixed"===s.options.strategy),popper:m(n)},s.reset=!1,s.placement=s.options.placement,s.orderedModifiers.forEach((function(e){return s.modifiersData[e.name]=Object.assign({},e.data)}));for(var r=0;r<s.orderedModifiers.length;r++)if(!0!==s.reset){var o=s.orderedModifiers[r],i=o.fn,a=o.options,c=void 0===a?{}:a,u=o.name;"function"==typeof i&&(s=i({state:s,options:c,name:u,instance:f})||s)}else s.reset=!1,r=-1}}},update:(o=function(){return new Promise((function(e){f.forceUpdate(),e(s)}))},function(){return c||(c=new Promise((function(e){Promise.resolve().then((function(){c=void 0,e(o())}))}))),c}),destroy:function(){d(),p=!0}};if(!P(e,t))return f;function d(){u.forEach((function(e){return e()})),u=[]}return f.setOptions(n).then((function(e){!p&&n.onFirstUpdate&&n.onFirstUpdate(e)})),f}}var R={passive:!0};const N={name:"eventListeners",enabled:!0,phase:"write",fn:function(){},effect:function(e){var t=e.state,n=e.instance,r=e.options,i=r.scroll,a=void 0===i||i,c=r.resize,s=void 0===c||c,u=o(t.elements.popper),p=[].concat(t.scrollParents.reference,t.scrollParents.popper);return a&&p.forEach((function(e){e.addEventListener("scroll",n.update,R)})),s&&u.addEventListener("resize",n.update,R),function(){a&&p.forEach((function(e){e.removeEventListener("scroll",n.update,R)})),s&&u.removeEventListener("resize",n.update,R)}},data:{}};function V(e){return e.split("-")[0]}function H(e){return e.split("-")[1]}function U(e){return["top","bottom"].indexOf(e)>=0?"x":"y"}function W(e){var t,n=e.reference,r=e.element,o=e.placement,i=o?V(o):null,a=o?H(o):null,c=n.x+n.width/2-r.width/2,s=n.y+n.height/2-r.height/2;switch(i){case k:t={x:c,y:n.y-r.height};break;case O:t={x:c,y:n.y+n.height};break;case j:t={x:n.x+n.width,y:s};break;case E:t={x:n.x-r.width,y:s};break;default:t={x:n.x,y:n.y}}var u=i?U(i):null;if(null!=u){var p="y"===u?"height":"width";switch(a){case T:t[u]=t[u]-(n[p]/2-r[p]/2);break;case S:t[u]=t[u]+(n[p]/2-r[p]/2)}}return t}const q={name:"popperOffsets",enabled:!0,phase:"read",fn:function(e){var t=e.state,n=e.name;t.modifiersData[n]=W({reference:t.rects.reference,element:t.rects.popper,strategy:"absolute",placement:t.placement})},data:{}};var F=Math.max,K=Math.min,J=Math.round,$={top:"auto",right:"auto",bottom:"auto",left:"auto"};function G(e){var t,n=e.popper,r=e.popperRect,i=e.placement,a=e.offsets,c=e.position,s=e.gpuAcceleration,u=e.adaptive,f=e.roundOffsets,l=!0===f?function(e){var t=e.x,n=e.y,r=window.devicePixelRatio||1;return{x:J(J(t*r)/r)||0,y:J(J(n*r)/r)||0}}(a):"function"==typeof f?f(a):a,v=l.x,m=void 0===v?0:v,h=l.y,g=void 0===h?0:h,b=a.hasOwnProperty("x"),y=a.hasOwnProperty("y"),w=E,C=k,A=window;if(u){var T=x(n),S="clientHeight",D="clientWidth";T===o(n)&&"static"!==d(T=p(n)).position&&(S="scrollHeight",D="scrollWidth"),T=T,i===k&&(C=O,g-=T[S]-r.height,g*=s?1:-1),i===E&&(w=j,m-=T[D]-r.width,m*=s?1:-1)}var L,_=Object.assign({position:c},u&&$);return s?Object.assign({},_,((L={})[C]=y?"0":"",L[w]=b?"0":"",L.transform=(A.devicePixelRatio||1)<2?"translate("+m+"px, "+g+"px)":"translate3d("+m+"px, "+g+"px, 0)",L)):Object.assign({},_,((t={})[C]=y?g+"px":"",t[w]=b?m+"px":"",t.transform="",t))}const Z={name:"applyStyles",enabled:!0,phase:"write",fn:function(e){var t=e.state;Object.keys(t.elements).forEach((function(e){var n=t.styles[e]||{},r=t.attributes[e]||{},o=t.elements[e];c(o)&&u(o)&&(Object.assign(o.style,n),Object.keys(r).forEach((function(e){var t=r[e];!1===t?o.removeAttribute(e):o.setAttribute(e,!0===t?"":t)})))}))},effect:function(e){var t=e.state,n={popper:{position:t.options.strategy,left:"0",top:"0",margin:"0"},arrow:{position:"absolute"},reference:{}};return Object.assign(t.elements.popper.style,n.popper),t.styles=n,t.elements.arrow&&Object.assign(t.elements.arrow.style,n.arrow),function(){Object.keys(t.elements).forEach((function(e){var r=t.elements[e],o=t.attributes[e]||{},i=Object.keys(t.styles.hasOwnProperty(e)?t.styles[e]:n[e]).reduce((function(e,t){return e[t]="",e}),{});c(r)&&u(r)&&(Object.assign(r.style,i),Object.keys(o).forEach((function(e){r.removeAttribute(e)})))}))}},requires:["computeStyles"]};var X={left:"right",right:"left",bottom:"top",top:"bottom"};function Y(e){return e.replace(/left|right|bottom|top/g,(function(e){return X[e]}))}var ee={start:"end",end:"start"};function te(e){return e.replace(/start|end/g,(function(e){return ee[e]}))}function ne(e,t){var n=t.getRootNode&&t.getRootNode();if(e.contains(t))return!0;if(n&&s(n)){var r=t;do{if(r&&e.isSameNode(r))return!0;r=r.parentNode||r.host}while(r)}return!1}function re(e){return Object.assign({},e,{left:e.x,top:e.y,right:e.x+e.width,bottom:e.y+e.height})}function oe(e,t){return t===D?re(function(e){var t=o(e),n=p(e),r=t.visualViewport,i=n.clientWidth,a=n.clientHeight,c=0,s=0;return r&&(i=r.width,a=r.height,/^((?!chrome|android).)*safari/i.test(navigator.userAgent)||(c=r.offsetLeft,s=r.offsetTop)),{width:i,height:a,x:c+f(e),y:s}}(e)):c(t)?function(e){var t=r(e);return t.top=t.top+e.clientTop,t.left=t.left+e.clientLeft,t.bottom=t.top+e.clientHeight,t.right=t.left+e.clientWidth,t.width=e.clientWidth,t.height=e.clientHeight,t.x=t.left,t.y=t.top,t}(t):re(function(e){var t,n=p(e),r=i(e),o=null==(t=e.ownerDocument)?void 0:t.body,a=F(n.scrollWidth,n.clientWidth,o?o.scrollWidth:0,o?o.clientWidth:0),c=F(n.scrollHeight,n.clientHeight,o?o.scrollHeight:0,o?o.clientHeight:0),s=-r.scrollLeft+f(e),u=-r.scrollTop;return"rtl"===d(o||n).direction&&(s+=F(n.clientWidth,o?o.clientWidth:0)-a),{width:a,height:c,x:s,y:u}}(p(e)))}function ie(e,t,n){var r="clippingParents"===t?function(e){var t=b(h(e)),n=["absolute","fixed"].indexOf(d(e).position)>=0&&c(e)?x(e):e;return a(n)?t.filter((function(e){return a(e)&&ne(e,n)&&"body"!==u(e)})):[]}(e):[].concat(t),o=[].concat(r,[n]),i=o[0],s=o.reduce((function(t,n){var r=oe(e,n);return t.top=F(r.top,t.top),t.right=K(r.right,t.right),t.bottom=K(r.bottom,t.bottom),t.left=F(r.left,t.left),t}),oe(e,i));return s.width=s.right-s.left,s.height=s.bottom-s.top,s.x=s.left,s.y=s.top,s}function ae(e){return Object.assign({},{top:0,right:0,bottom:0,left:0},e)}function ce(e,t){return t.reduce((function(t,n){return t[n]=e,t}),{})}function se(e,t){void 0===t&&(t={});var n=t,o=n.placement,i=void 0===o?e.placement:o,c=n.boundary,s=void 0===c?"clippingParents":c,u=n.rootBoundary,f=void 0===u?D:u,d=n.elementContext,l=void 0===d?L:d,v=n.altBoundary,m=void 0!==v&&v,h=n.padding,g=void 0===h?0:h,b=ae("number"!=typeof g?g:ce(g,A)),y=l===L?"reference":L,w=e.elements.reference,x=e.rects.popper,E=e.elements[m?y:l],C=ie(a(E)?E:E.contextElement||p(e.elements.popper),s,f),T=r(w),S=W({reference:T,element:x,strategy:"absolute",placement:i}),_=re(Object.assign({},x,S)),I=l===L?_:T,M={top:C.top-I.top+b.top,bottom:I.bottom-C.bottom+b.bottom,left:C.left-I.left+b.left,right:I.right-C.right+b.right},Q=e.modifiersData.offset;if(l===L&&Q){var z=Q[i];Object.keys(M).forEach((function(e){var t=[j,O].indexOf(e)>=0?1:-1,n=[k,O].indexOf(e)>=0?"y":"x";M[e]+=z[n]*t}))}return M}function ue(e,t,n){return F(e,K(t,n))}function pe(e,t,n){return void 0===n&&(n={x:0,y:0}),{top:e.top-t.height-n.y,right:e.right-t.width+n.x,bottom:e.bottom-t.height+n.y,left:e.left-t.width-n.x}}function fe(e){return[k,j,O,E].some((function(t){return e[t]>=0}))}var de=B({defaultModifiers:[N,q,{name:"computeStyles",enabled:!0,phase:"beforeWrite",fn:function(e){var t=e.state,n=e.options,r=n.gpuAcceleration,o=void 0===r||r,i=n.adaptive,a=void 0===i||i,c=n.roundOffsets,s=void 0===c||c,u={placement:V(t.placement),popper:t.elements.popper,popperRect:t.rects.popper,gpuAcceleration:o};null!=t.modifiersData.popperOffsets&&(t.styles.popper=Object.assign({},t.styles.popper,G(Object.assign({},u,{offsets:t.modifiersData.popperOffsets,position:t.options.strategy,adaptive:a,roundOffsets:s})))),null!=t.modifiersData.arrow&&(t.styles.arrow=Object.assign({},t.styles.arrow,G(Object.assign({},u,{offsets:t.modifiersData.arrow,position:"absolute",adaptive:!1,roundOffsets:s})))),t.attributes.popper=Object.assign({},t.attributes.popper,{"data-popper-placement":t.placement})},data:{}},Z,{name:"offset",enabled:!0,phase:"main",requires:["popperOffsets"],fn:function(e){var t=e.state,n=e.options,r=e.name,o=n.offset,i=void 0===o?[0,0]:o,a=I.reduce((function(e,n){return e[n]=function(e,t,n){var r=V(e),o=[E,k].indexOf(r)>=0?-1:1,i="function"==typeof n?n(Object.assign({},t,{placement:e})):n,a=i[0],c=i[1];return a=a||0,c=(c||0)*o,[E,j].indexOf(r)>=0?{x:c,y:a}:{x:a,y:c}}(n,t.rects,i),e}),{}),c=a[t.placement],s=c.x,u=c.y;null!=t.modifiersData.popperOffsets&&(t.modifiersData.popperOffsets.x+=s,t.modifiersData.popperOffsets.y+=u),t.modifiersData[r]=a}},{name:"flip",enabled:!0,phase:"main",fn:function(e){var t=e.state,n=e.options,r=e.name;if(!t.modifiersData[r]._skip){for(var o=n.mainAxis,i=void 0===o||o,a=n.altAxis,c=void 0===a||a,s=n.fallbackPlacements,u=n.padding,p=n.boundary,f=n.rootBoundary,d=n.altBoundary,l=n.flipVariations,v=void 0===l||l,m=n.allowedAutoPlacements,h=t.options.placement,g=V(h),b=s||(g===h||!v?[Y(h)]:function(e){if(V(e)===C)return[];var t=Y(e);return[te(e),t,te(t)]}(h)),y=[h].concat(b).reduce((function(e,n){return e.concat(V(n)===C?function(e,t){void 0===t&&(t={});var n=t,r=n.placement,o=n.boundary,i=n.rootBoundary,a=n.padding,c=n.flipVariations,s=n.allowedAutoPlacements,u=void 0===s?I:s,p=H(r),f=p?c?_:_.filter((function(e){return H(e)===p})):A,d=f.filter((function(e){return u.indexOf(e)>=0}));0===d.length&&(d=f);var l=d.reduce((function(t,n){return t[n]=se(e,{placement:n,boundary:o,rootBoundary:i,padding:a})[V(n)],t}),{});return Object.keys(l).sort((function(e,t){return l[e]-l[t]}))}(t,{placement:n,boundary:p,rootBoundary:f,padding:u,flipVariations:v,allowedAutoPlacements:m}):n)}),[]),w=t.rects.reference,x=t.rects.popper,S=new Map,D=!0,L=y[0],M=0;M<y.length;M++){var Q=y[M],z=V(Q),P=H(Q)===T,B=[k,O].indexOf(z)>=0,R=B?"width":"height",N=se(t,{placement:Q,boundary:p,rootBoundary:f,altBoundary:d,padding:u}),U=B?P?j:E:P?O:k;w[R]>x[R]&&(U=Y(U));var W=Y(U),q=[];if(i&&q.push(N[z]<=0),c&&q.push(N[U]<=0,N[W]<=0),q.every((function(e){return e}))){L=Q,D=!1;break}S.set(Q,q)}if(D)for(var F=function(e){var t=y.find((function(t){var n=S.get(t);if(n)return n.slice(0,e).every((function(e){return e}))}));if(t)return L=t,"break"},K=v?3:1;K>0;K--){if("break"===F(K))break}t.placement!==L&&(t.modifiersData[r]._skip=!0,t.placement=L,t.reset=!0)}},requiresIfExists:["offset"],data:{_skip:!1}},{name:"preventOverflow",enabled:!0,phase:"main",fn:function(e){var t=e.state,n=e.options,r=e.name,o=n.mainAxis,i=void 0===o||o,a=n.altAxis,c=void 0!==a&&a,s=n.boundary,u=n.rootBoundary,p=n.altBoundary,f=n.padding,d=n.tether,l=void 0===d||d,v=n.tetherOffset,h=void 0===v?0:v,g=se(t,{boundary:s,rootBoundary:u,padding:f,altBoundary:p}),b=V(t.placement),y=H(t.placement),w=!y,C=U(b),A="x"===C?"y":"x",S=t.modifiersData.popperOffsets,D=t.rects.reference,L=t.rects.popper,_="function"==typeof h?h(Object.assign({},t.rects,{placement:t.placement})):h,I={x:0,y:0};if(S){if(i||c){var M="y"===C?k:E,Q="y"===C?O:j,z="y"===C?"height":"width",P=S[C],B=S[C]+g[M],R=S[C]-g[Q],N=l?-L[z]/2:0,W=y===T?D[z]:L[z],q=y===T?-L[z]:-D[z],J=t.elements.arrow,$=l&&J?m(J):{width:0,height:0},G=t.modifiersData["arrow#persistent"]?t.modifiersData["arrow#persistent"].padding:{top:0,right:0,bottom:0,left:0},Z=G[M],X=G[Q],Y=ue(0,D[z],$[z]),ee=w?D[z]/2-N-Y-Z-_:W-Y-Z-_,te=w?-D[z]/2+N+Y+X+_:q+Y+X+_,ne=t.elements.arrow&&x(t.elements.arrow),re=ne?"y"===C?ne.clientTop||0:ne.clientLeft||0:0,oe=t.modifiersData.offset?t.modifiersData.offset[t.placement][C]:0,ie=S[C]+ee-oe-re,ae=S[C]+te-oe;if(i){var ce=ue(l?K(B,ie):B,P,l?F(R,ae):R);S[C]=ce,I[C]=ce-P}if(c){var pe="x"===C?k:E,fe="x"===C?O:j,de=S[A],le=de+g[pe],ve=de-g[fe],me=ue(l?K(le,ie):le,de,l?F(ve,ae):ve);S[A]=me,I[A]=me-de}}t.modifiersData[r]=I}},requiresIfExists:["offset"]},{name:"arrow",enabled:!0,phase:"main",fn:function(e){var t,n=e.state,r=e.name,o=e.options,i=n.elements.arrow,a=n.modifiersData.popperOffsets,c=V(n.placement),s=U(c),u=[E,j].indexOf(c)>=0?"height":"width";if(i&&a){var p=function(e,t){return ae("number"!=typeof(e="function"==typeof e?e(Object.assign({},t.rects,{placement:t.placement})):e)?e:ce(e,A))}(o.padding,n),f=m(i),d="y"===s?k:E,l="y"===s?O:j,v=n.rects.reference[u]+n.rects.reference[s]-a[s]-n.rects.popper[u],h=a[s]-n.rects.reference[s],g=x(i),b=g?"y"===s?g.clientHeight||0:g.clientWidth||0:0,y=v/2-h/2,w=p[d],C=b-f[u]-p[l],T=b/2-f[u]/2+y,S=ue(w,T,C),D=s;n.modifiersData[r]=((t={})[D]=S,t.centerOffset=S-T,t)}},effect:function(e){var t=e.state,n=e.options.element,r=void 0===n?"[data-popper-arrow]":n;null!=r&&("string"!=typeof r||(r=t.elements.popper.querySelector(r)))&&ne(t.elements.popper,r)&&(t.elements.arrow=r)},requires:["popperOffsets"],requiresIfExists:["preventOverflow"]},{name:"hide",enabled:!0,phase:"main",requiresIfExists:["preventOverflow"],fn:function(e){var t=e.state,n=e.name,r=t.rects.reference,o=t.rects.popper,i=t.modifiersData.preventOverflow,a=se(t,{elementContext:"reference"}),c=se(t,{altBoundary:!0}),s=pe(a,r),u=pe(c,o,i),p=fe(s),f=fe(u);t.modifiersData[n]={referenceClippingOffsets:s,popperEscapeOffsets:u,isReferenceHidden:p,hasPopperEscaped:f},t.attributes.popper=Object.assign({},t.attributes.popper,{"data-popper-reference-hidden":p,"data-popper-escaped":f})}}]}),le="tippy-content",ve="tippy-backdrop",me="tippy-arrow",he="tippy-svg-arrow",ge={passive:!0,capture:!0};function be(e,t,n){if(Array.isArray(e)){var r=e[t];return null==r?Array.isArray(n)?n[t]:n:r}return e}function ye(e,t){var n={}.toString.call(e);return 0===n.indexOf("[object")&&n.indexOf(t+"]")>-1}function we(e,t){return"function"==typeof e?e.apply(void 0,t):e}function xe(e,t){return 0===t?e:function(r){clearTimeout(n),n=setTimeout((function(){e(r)}),t)};var n}function ke(e){return[].concat(e)}function Oe(e,t){-1===e.indexOf(t)&&e.push(t)}function je(e){return e.split("-")[0]}function Ee(e){return[].slice.call(e)}function Ce(){return document.createElement("div")}function Ae(e){return["Element","Fragment"].some((function(t){return ye(e,t)}))}function Te(e){return ye(e,"MouseEvent")}function Se(e){return!(!e||!e._tippy||e._tippy.reference!==e)}function De(e){return Ae(e)?[e]:function(e){return ye(e,"NodeList")}(e)?Ee(e):Array.isArray(e)?e:Ee(document.querySelectorAll(e))}function Le(e,t){e.forEach((function(e){e&&(e.style.transitionDuration=t+"ms")}))}function _e(e,t){e.forEach((function(e){e&&e.setAttribute("data-state",t)}))}function Ie(e){var t,n=ke(e)[0];return(null==n||null==(t=n.ownerDocument)?void 0:t.body)?n.ownerDocument:document}function Me(e,t,n){var r=t+"EventListener";["transitionend","webkitTransitionEnd"].forEach((function(t){e[r](t,n)}))}var Qe={isTouch:!1},ze=0;function Pe(){Qe.isTouch||(Qe.isTouch=!0,window.performance&&document.addEventListener("mousemove",Be))}function Be(){var e=performance.now();e-ze<20&&(Qe.isTouch=!1,document.removeEventListener("mousemove",Be)),ze=e}function Re(){var e=document.activeElement;if(Se(e)){var t=e._tippy;e.blur&&!t.state.isVisible&&e.blur()}}var Ne="undefined"!=typeof window&&"undefined"!=typeof document?navigator.userAgent:"",Ve=/MSIE |Trident\//.test(Ne);var He={animateFill:!1,followCursor:!1,inlinePositioning:!1,sticky:!1},Ue=Object.assign({appendTo:function(){return document.body},aria:{content:"auto",expanded:"auto"},delay:0,duration:[300,250],getReferenceClientRect:null,hideOnClick:!0,ignoreAttributes:!1,interactive:!1,interactiveBorder:2,interactiveDebounce:0,moveTransition:"",offset:[0,10],onAfterUpdate:function(){},onBeforeUpdate:function(){},onCreate:function(){},onDestroy:function(){},onHidden:function(){},onHide:function(){},onMount:function(){},onShow:function(){},onShown:function(){},onTrigger:function(){},onUntrigger:function(){},onClickOutside:function(){},placement:"top",plugins:[],popperOptions:{},render:null,showOnCreate:!1,touch:!0,trigger:"mouseenter focus",triggerTarget:null},He,{},{allowHTML:!1,animation:"fade",arrow:!0,content:"",inertia:!1,maxWidth:350,role:"tooltip",theme:"",zIndex:9999}),We=Object.keys(Ue);function qe(e){var t=(e.plugins||[]).reduce((function(t,n){var r=n.name,o=n.defaultValue;return r&&(t[r]=void 0!==e[r]?e[r]:o),t}),{});return Object.assign({},e,{},t)}function Fe(e,t){var n=Object.assign({},t,{content:we(t.content,[e])},t.ignoreAttributes?{}:function(e,t){return(t?Object.keys(qe(Object.assign({},Ue,{plugins:t}))):We).reduce((function(t,n){var r=(e.getAttribute("data-tippy-"+n)||"").trim();if(!r)return t;if("content"===n)t[n]=r;else try{t[n]=JSON.parse(r)}catch(e){t[n]=r}return t}),{})}(e,t.plugins));return n.aria=Object.assign({},Ue.aria,{},n.aria),n.aria={expanded:"auto"===n.aria.expanded?t.interactive:n.aria.expanded,content:"auto"===n.aria.content?t.interactive?null:"describedby":n.aria.content},n}function Ke(e,t){e.innerHTML=t}function Je(e){var t=Ce();return!0===e?t.className=me:(t.className=he,Ae(e)?t.appendChild(e):Ke(t,e)),t}function $e(e,t){Ae(t.content)?(Ke(e,""),e.appendChild(t.content)):"function"!=typeof t.content&&(t.allowHTML?Ke(e,t.content):e.textContent=t.content)}function Ge(e){var t=e.firstElementChild,n=Ee(t.children);return{box:t,content:n.find((function(e){return e.classList.contains(le)})),arrow:n.find((function(e){return e.classList.contains(me)||e.classList.contains(he)})),backdrop:n.find((function(e){return e.classList.contains(ve)}))}}function Ze(e){var t=Ce(),n=Ce();n.className="tippy-box",n.setAttribute("data-state","hidden"),n.setAttribute("tabindex","-1");var r=Ce();function o(n,r){var o=Ge(t),i=o.box,a=o.content,c=o.arrow;r.theme?i.setAttribute("data-theme",r.theme):i.removeAttribute("data-theme"),"string"==typeof r.animation?i.setAttribute("data-animation",r.animation):i.removeAttribute("data-animation"),r.inertia?i.setAttribute("data-inertia",""):i.removeAttribute("data-inertia"),i.style.maxWidth="number"==typeof r.maxWidth?r.maxWidth+"px":r.maxWidth,r.role?i.setAttribute("role",r.role):i.removeAttribute("role"),n.content===r.content&&n.allowHTML===r.allowHTML||$e(a,e.props),r.arrow?c?n.arrow!==r.arrow&&(i.removeChild(c),i.appendChild(Je(r.arrow))):i.appendChild(Je(r.arrow)):c&&i.removeChild(c)}return r.className=le,r.setAttribute("data-state","hidden"),$e(r,e.props),t.appendChild(n),n.appendChild(r),o(e.props,e.props),{popper:t,onUpdate:o}}Ze.$$tippy=!0;var Xe=1,Ye=[],et=[];function tt(e,t){var n,r,o,i,a,c,s,u,p,f=Fe(e,Object.assign({},Ue,{},qe((n=t,Object.keys(n).reduce((function(e,t){return void 0!==n[t]&&(e[t]=n[t]),e}),{}))))),d=!1,l=!1,v=!1,m=!1,h=[],g=xe(J,f.interactiveDebounce),b=Xe++,y=(p=f.plugins).filter((function(e,t){return p.indexOf(e)===t})),w={id:b,reference:e,popper:Ce(),popperInstance:null,props:f,state:{isEnabled:!0,isVisible:!1,isDestroyed:!1,isMounted:!1,isShown:!1},plugins:y,clearDelayTimeouts:function(){clearTimeout(r),clearTimeout(o),cancelAnimationFrame(i)},setProps:function(t){0;if(w.state.isDestroyed)return;M("onBeforeUpdate",[w,t]),F();var n=w.props,r=Fe(e,Object.assign({},w.props,{},t,{ignoreAttributes:!0}));w.props=r,q(),n.interactiveDebounce!==r.interactiveDebounce&&(P(),g=xe(J,r.interactiveDebounce));n.triggerTarget&&!r.triggerTarget?ke(n.triggerTarget).forEach((function(e){e.removeAttribute("aria-expanded")})):r.triggerTarget&&e.removeAttribute("aria-expanded");z(),I(),O&&O(n,r);w.popperInstance&&(X(),ee().forEach((function(e){requestAnimationFrame(e._tippy.popperInstance.forceUpdate)})));M("onAfterUpdate",[w,t])},setContent:function(e){w.setProps({content:e})},show:function(){0;var e=w.state.isVisible,t=w.state.isDestroyed,n=!w.state.isEnabled,r=Qe.isTouch&&!w.props.touch,o=be(w.props.duration,0,Ue.duration);if(e||t||n||r)return;if(S().hasAttribute("disabled"))return;if(M("onShow",[w],!1),!1===w.props.onShow(w))return;w.state.isVisible=!0,T()&&(k.style.visibility="visible");I(),V(),w.state.isMounted||(k.style.transition="none");if(T()){var i=L(),a=i.box,c=i.content;Le([a,c],0)}s=function(){var e;if(w.state.isVisible&&!m){if(m=!0,k.offsetHeight,k.style.transition=w.props.moveTransition,T()&&w.props.animation){var t=L(),n=t.box,r=t.content;Le([n,r],o),_e([n,r],"visible")}Q(),z(),Oe(et,w),null==(e=w.popperInstance)||e.forceUpdate(),w.state.isMounted=!0,M("onMount",[w]),w.props.animation&&T()&&function(e,t){U(e,t)}(o,(function(){w.state.isShown=!0,M("onShown",[w])}))}},function(){var e,t=w.props.appendTo,n=S();e=w.props.interactive&&t===Ue.appendTo||"parent"===t?n.parentNode:we(t,[n]);e.contains(k)||e.appendChild(k);X(),!1}()},hide:function(){0;var e=!w.state.isVisible,t=w.state.isDestroyed,n=!w.state.isEnabled,r=be(w.props.duration,1,Ue.duration);if(e||t||n)return;if(M("onHide",[w],!1),!1===w.props.onHide(w))return;w.state.isVisible=!1,w.state.isShown=!1,m=!1,d=!1,T()&&(k.style.visibility="hidden");if(P(),H(),I(),T()){var o=L(),i=o.box,a=o.content;w.props.animation&&(Le([i,a],r),_e([i,a],"hidden"))}Q(),z(),w.props.animation?T()&&function(e,t){U(e,(function(){!w.state.isVisible&&k.parentNode&&k.parentNode.contains(k)&&t()}))}(r,w.unmount):w.unmount()},hideWithInteractivity:function(e){0;D().addEventListener("mousemove",g),Oe(Ye,g),g(e)},enable:function(){w.state.isEnabled=!0},disable:function(){w.hide(),w.state.isEnabled=!1},unmount:function(){0;w.state.isVisible&&w.hide();if(!w.state.isMounted)return;Y(),ee().forEach((function(e){e._tippy.unmount()})),k.parentNode&&k.parentNode.removeChild(k);et=et.filter((function(e){return e!==w})),w.state.isMounted=!1,M("onHidden",[w])},destroy:function(){0;if(w.state.isDestroyed)return;w.clearDelayTimeouts(),w.unmount(),F(),delete e._tippy,w.state.isDestroyed=!0,M("onDestroy",[w])}};if(!f.render)return w;var x=f.render(w),k=x.popper,O=x.onUpdate;k.setAttribute("data-tippy-root",""),k.id="tippy-"+w.id,w.popper=k,e._tippy=w,k._tippy=w;var j=y.map((function(e){return e.fn(w)})),E=e.hasAttribute("aria-expanded");return q(),z(),I(),M("onCreate",[w]),f.showOnCreate&&te(),k.addEventListener("mouseenter",(function(){w.props.interactive&&w.state.isVisible&&w.clearDelayTimeouts()})),k.addEventListener("mouseleave",(function(e){w.props.interactive&&w.props.trigger.indexOf("mouseenter")>=0&&(D().addEventListener("mousemove",g),g(e))})),w;function C(){var e=w.props.touch;return Array.isArray(e)?e:[e,0]}function A(){return"hold"===C()[0]}function T(){var e;return!!(null==(e=w.props.render)?void 0:e.$$tippy)}function S(){return u||e}function D(){var e=S().parentNode;return e?Ie(e):document}function L(){return Ge(k)}function _(e){return w.state.isMounted&&!w.state.isVisible||Qe.isTouch||a&&"focus"===a.type?0:be(w.props.delay,e?0:1,Ue.delay)}function I(){k.style.pointerEvents=w.props.interactive&&w.state.isVisible?"":"none",k.style.zIndex=""+w.props.zIndex}function M(e,t,n){var r;(void 0===n&&(n=!0),j.forEach((function(n){n[e]&&n[e].apply(void 0,t)})),n)&&(r=w.props)[e].apply(r,t)}function Q(){var t=w.props.aria;if(t.content){var n="aria-"+t.content,r=k.id;ke(w.props.triggerTarget||e).forEach((function(e){var t=e.getAttribute(n);if(w.state.isVisible)e.setAttribute(n,t?t+" "+r:r);else{var o=t&&t.replace(r,"").trim();o?e.setAttribute(n,o):e.removeAttribute(n)}}))}}function z(){!E&&w.props.aria.expanded&&ke(w.props.triggerTarget||e).forEach((function(e){w.props.interactive?e.setAttribute("aria-expanded",w.state.isVisible&&e===S()?"true":"false"):e.removeAttribute("aria-expanded")}))}function P(){D().removeEventListener("mousemove",g),Ye=Ye.filter((function(e){return e!==g}))}function B(e){if(!(Qe.isTouch&&(v||"mousedown"===e.type)||w.props.interactive&&k.contains(e.target))){if(S().contains(e.target)){if(Qe.isTouch)return;if(w.state.isVisible&&w.props.trigger.indexOf("click")>=0)return}else M("onClickOutside",[w,e]);!0===w.props.hideOnClick&&(w.clearDelayTimeouts(),w.hide(),l=!0,setTimeout((function(){l=!1})),w.state.isMounted||H())}}function R(){v=!0}function N(){v=!1}function V(){var e=D();e.addEventListener("mousedown",B,!0),e.addEventListener("touchend",B,ge),e.addEventListener("touchstart",N,ge),e.addEventListener("touchmove",R,ge)}function H(){var e=D();e.removeEventListener("mousedown",B,!0),e.removeEventListener("touchend",B,ge),e.removeEventListener("touchstart",N,ge),e.removeEventListener("touchmove",R,ge)}function U(e,t){var n=L().box;function r(e){e.target===n&&(Me(n,"remove",r),t())}if(0===e)return t();Me(n,"remove",c),Me(n,"add",r),c=r}function W(t,n,r){void 0===r&&(r=!1),ke(w.props.triggerTarget||e).forEach((function(e){e.addEventListener(t,n,r),h.push({node:e,eventType:t,handler:n,options:r})}))}function q(){var e;A()&&(W("touchstart",K,{passive:!0}),W("touchend",$,{passive:!0})),(e=w.props.trigger,e.split(/\s+/).filter(Boolean)).forEach((function(e){if("manual"!==e)switch(W(e,K),e){case"mouseenter":W("mouseleave",$);break;case"focus":W(Ve?"focusout":"blur",G);break;case"focusin":W("focusout",G)}}))}function F(){h.forEach((function(e){var t=e.node,n=e.eventType,r=e.handler,o=e.options;t.removeEventListener(n,r,o)})),h=[]}function K(e){var t,n=!1;if(w.state.isEnabled&&!Z(e)&&!l){var r="focus"===(null==(t=a)?void 0:t.type);a=e,u=e.currentTarget,z(),!w.state.isVisible&&Te(e)&&Ye.forEach((function(t){return t(e)})),"click"===e.type&&(w.props.trigger.indexOf("mouseenter")<0||d)&&!1!==w.props.hideOnClick&&w.state.isVisible?n=!0:te(e),"click"===e.type&&(d=!n),n&&!r&&ne(e)}}function J(e){var t=e.target,n=S().contains(t)||k.contains(t);"mousemove"===e.type&&n||function(e,t){var n=t.clientX,r=t.clientY;return e.every((function(e){var t=e.popperRect,o=e.popperState,i=e.props.interactiveBorder,a=je(o.placement),c=o.modifiersData.offset;if(!c)return!0;var s="bottom"===a?c.top.y:0,u="top"===a?c.bottom.y:0,p="right"===a?c.left.x:0,f="left"===a?c.right.x:0,d=t.top-r+s>i,l=r-t.bottom-u>i,v=t.left-n+p>i,m=n-t.right-f>i;return d||l||v||m}))}(ee().concat(k).map((function(e){var t,n=null==(t=e._tippy.popperInstance)?void 0:t.state;return n?{popperRect:e.getBoundingClientRect(),popperState:n,props:f}:null})).filter(Boolean),e)&&(P(),ne(e))}function $(e){Z(e)||w.props.trigger.indexOf("click")>=0&&d||(w.props.interactive?w.hideWithInteractivity(e):ne(e))}function G(e){w.props.trigger.indexOf("focusin")<0&&e.target!==S()||w.props.interactive&&e.relatedTarget&&k.contains(e.relatedTarget)||ne(e)}function Z(e){return!!Qe.isTouch&&A()!==e.type.indexOf("touch")>=0}function X(){Y();var t=w.props,n=t.popperOptions,r=t.placement,o=t.offset,i=t.getReferenceClientRect,a=t.moveTransition,c=T()?Ge(k).arrow:null,u=i?{getBoundingClientRect:i,contextElement:i.contextElement||S()}:e,p=[{name:"offset",options:{offset:o}},{name:"preventOverflow",options:{padding:{top:2,bottom:2,left:5,right:5}}},{name:"flip",options:{padding:5}},{name:"computeStyles",options:{adaptive:!a}},{name:"$$tippy",enabled:!0,phase:"beforeWrite",requires:["computeStyles"],fn:function(e){var t=e.state;if(T()){var n=L().box;["placement","reference-hidden","escaped"].forEach((function(e){"placement"===e?n.setAttribute("data-placement",t.placement):t.attributes.popper["data-popper-"+e]?n.setAttribute("data-"+e,""):n.removeAttribute("data-"+e)})),t.attributes.popper={}}}}];T()&&c&&p.push({name:"arrow",options:{element:c,padding:3}}),p.push.apply(p,(null==n?void 0:n.modifiers)||[]),w.popperInstance=de(u,k,Object.assign({},n,{placement:r,onFirstUpdate:s,modifiers:p}))}function Y(){w.popperInstance&&(w.popperInstance.destroy(),w.popperInstance=null)}function ee(){return Ee(k.querySelectorAll("[data-tippy-root]"))}function te(e){w.clearDelayTimeouts(),e&&M("onTrigger",[w,e]),V();var t=_(!0),n=C(),o=n[0],i=n[1];Qe.isTouch&&"hold"===o&&i&&(t=i),t?r=setTimeout((function(){w.show()}),t):w.show()}function ne(e){if(w.clearDelayTimeouts(),M("onUntrigger",[w,e]),w.state.isVisible){if(!(w.props.trigger.indexOf("mouseenter")>=0&&w.props.trigger.indexOf("click")>=0&&["mouseleave","mousemove"].indexOf(e.type)>=0&&d)){var t=_(!1);t?o=setTimeout((function(){w.state.isVisible&&w.hide()}),t):i=requestAnimationFrame((function(){w.hide()}))}}else H()}}function nt(e,t){void 0===t&&(t={});var n=Ue.plugins.concat(t.plugins||[]);document.addEventListener("touchstart",Pe,ge),window.addEventListener("blur",Re);var r=Object.assign({},t,{plugins:n}),o=De(e).reduce((function(e,t){var n=t&&tt(t,r);return n&&e.push(n),e}),[]);return Ae(e)?o[0]:o}nt.defaultProps=Ue,nt.setDefaultProps=function(e){Object.keys(e).forEach((function(t){Ue[t]=e[t]}))},nt.currentInput=Qe;Object.assign({},Z,{effect:function(e){var t=e.state,n={popper:{position:t.options.strategy,left:"0",top:"0",margin:"0"},arrow:{position:"absolute"},reference:{}};Object.assign(t.elements.popper.style,n.popper),t.styles=n,t.elements.arrow&&Object.assign(t.elements.arrow.style,n.arrow)}});nt.setDefaultProps({render:Ze});const rt=nt}},t={};function n(r){var o=t[r];if(void 0!==o)return o.exports;var i=t[r]={id:r,exports:{}};return e[r](i,i.exports,n),i.exports}n.n=e=>{var t=e&&e.__esModule?()=>e.default:()=>e;return n.d(t,{a:t}),t},n.d=(e,t)=>{for(var r in t)n.o(t,r)&&!n.o(e,r)&&Object.defineProperty(e,r,{enumerable:!0,get:t[r]})},n.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t),n(108),n(615),n(502);n(861)})();
+(()=>{var e={278:(e,t,n)=>{"use strict";var r={readonly:function(e){var t=45,n=46,r=97,o=99,i=118,a=120,c=121,s=122,u=e.which;if(e.ctrlKey&&u==c||e.ctrlKey&&u==s||e.ctrlKey&&u==a||e.ctrlKey&&u==i||e.ctrlKey&&u==o||e.shiftKey&&e.keyCode==t||e.shiftKey&&e.keyCode==n||e.ctrlKey&&e.keyCode==t||e.ctrlKey&&u==r)return 0;var p=e||window.event,f=p.keyCode||p.which;f=String.fromCharCode(f);/[]|\./.test(f)||(p.returnValue=!1,p.preventDefault())},isJson:function(e){try{return JSON.parse(e)}catch(e){}},copyText:function(){var e=jQuery(this),t=e.text(),n=jQuery("<input>");jQuery("body").append(n),n.val(t).select(),document.execCommand("copy"),n.remove(),e.addClass("copied"),setTimeout((function(){return e.removeClass("copied")}),500)},alert:function(e){function t(t){return e.apply(this,arguments)}return t.toString=function(){return e.toString()},t}((function(e){alert(e)})),confirm:function(e){return window.confirm(e)}};var o=function(e){return function(e){jQuery.ajax({type:"POST",url:ajaxurl,data:{action:"ccfw_cookie_store",payload:e},beforeSend:function(){jQuery(".ccfw-ajax").addClass("active").fadeIn("fast")}}).done((function(){jQuery(".ccfw-ajax").fadeOut("fast",(function(){jQuery(this).removeClass("active")})),A.sectionsStored=!0,jQuery("#submit").click()})).fail((function(e){console.log(e)}))}(e)},i=function(){jQuery.ajax({url:ajaxurl,data:{action:"ccfw_cookie_get"}}).done((function(e){A.sections=JSON.parse(e)||{},Object.keys(A.sections).length>0&&(A.sectionsLoaded=!0),A.debug.output()}))};function a(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=null==e?null:"undefined"!=typeof Symbol&&e[Symbol.iterator]||e["@@iterator"];if(null==n)return;var r,o,i=[],a=!0,c=!1;try{for(n=n.call(e);!(a=(r=n.next()).done)&&(i.push(r.value),!t||i.length!==t);a=!0);}catch(e){c=!0,o=e}finally{try{a||null==n.return||n.return()}finally{if(c)throw o}}return i}(e,t)||function(e,t){if(!e)return;if("string"==typeof e)return c(e,t);var n=Object.prototype.toString.call(e).slice(8,-1);"Object"===n&&e.constructor&&(n=e.constructor.name);if("Map"===n||"Set"===n)return Array.from(e);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return c(e,t)}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function c(e,t){(null==t||t>e.length)&&(t=e.length);for(var n=0,r=new Array(t);n<t;n++)r[n]=e[n];return r}var s={init:function(){_.cookies.get();var e=setInterval((function(){A.sectionsLoaded&&(clearInterval(e),s.load())}),100)},load:function(){A.building=!0;for(var e=0,t=Object.entries(A.sections);e<t.length;e++){var n=a(t[e],2),r=n[0],o=n[1];if(A.create.section(r),o)for(var i=0,c=Object.entries(o);i<c.length;i++){var s=a(c[i],2),u=(s[0],s[1]);if(A.create.group({section:r,group:u.name,description:u.description,id:u.allowlistID}),u.cookies)for(var p=0,f=Object.entries(u.cookies);p<f.length;p++){var d=a(f[p],2),l=d[0],v=d[1];A.create.cookie(l.replace("row_",""),{section:r,group:u.name,row:v})}}}A.building=!1}};function u(){var e=jQuery(this).is(":checked"),t=jQuery("#"+A.debug.container);e?(A.debug.active=!0,0===t.length&&A.debug.make(),t.fadeIn()):(A.debug.active=!1,t.fadeOut())}var p=function(){var e=jQuery("."+A.debug.import.textarea).val(),t=r.isJson(e);t&&(A.sections=t,s.load(),A.debug.output())},f={save:function(e,t){return function(e,t){var n=jQuery(".ccfw-"+e+"-group-save");n.off("click",t),n.on("click",t)}(e,t)},remove:function(e){return function(e){var t=jQuery(".ccfw-group-remove");t.off("click",e),t.on("click",e)}(e)},allowlist:{save:function(e){return function(e){jQuery(".ccfw-cookie-row__gtm-allowlist-id").on("keyup",e)}(e)}},description:{save:function(e){return function(e){jQuery(".ccfw-group__description input").on("keyup",e)}(e)}}},d={add:function(e){return function(e){var t=jQuery(".ccfw-cookie-row-add");t.off("click",e),t.on("click",e)}(e)},save:function(e){return function(e){var t=jQuery(".ccfw-cookie-row input");t.off("keyup",e),t.on("keyup",e)}(e)},remove:function(e){return function(e){var t=jQuery(".ccfw-cookie-row-remove");t.off("click",e),t.on("click",e)}(e)}},l={submit:function(e){return jQuery("#ccfw-form").on("submit",e)},input:{focusBlur:function(e){return jQuery("#"+A.appContainer).on("focus blur","input",e)}}},v={toggle:function(){jQuery("#"+A.debug.checkbox).on("change",u)},import:function(){return(e=jQuery("."+A.debug.import.textarea)).on("keypress",r.readonly),void e.on("keyup",p);var e},copy:function(){jQuery("#"+A.debug.preContainer).find("pre").on("click",r.copyText)}},m={add:'<path d="M24 10h-10v-10h-4v10h-10v4h10v10h4v-10h10z"/>',success:'<path d="M21.856 10.303c.086.554.144 1.118.144 1.697 0 6.075-4.925 11-11 11s-11-4.925-11-11 4.925-11 11-11c2.347 0 4.518.741 6.304 1.993l-1.422 1.457c-1.408-.913-3.082-1.45-4.882-1.45-4.962 0-9 4.038-9 9s4.038 9 9 9c4.894 0 8.879-3.928 8.99-8.795l1.866-1.902zm-.952-8.136l-9.404 9.639-3.843-3.614-3.095 3.098 6.938 6.71 12.5-12.737-3.096-3.096z"/>',check:'<path d="M9 21.035l-9-8.638 2.791-2.87 6.156 5.874 12.21-12.436 2.843 2.817z"/>',cross:'<path d="M23.954 21.03l-9.184-9.095 9.092-9.174-2.832-2.807-9.09 9.179-9.176-9.088-2.81 2.81 9.186 9.105-9.095 9.184 2.81 2.81 9.112-9.192 9.18 9.1z"/>',crossArrow:'<path d="M7 5h17v16h-17l-7-7.972 7-8.028zm7 6.586l-2.586-2.586-1.414 1.414 2.586 2.586-2.586 2.586 1.414 1.414 2.586-2.586 2.586 2.586 1.414-1.414-2.586-2.586 2.586-2.586-1.414-1.414-2.586 2.586z"/>',error:'<path d="M16.142 2l5.858 5.858v8.284l-5.858 5.858h-8.284l-5.858-5.858v-8.284l5.858-5.858h8.284zm.829-2h-9.942l-7.029 7.029v9.941l7.029 7.03h9.941l7.03-7.029v-9.942l-7.029-7.029zm-8.482 16.992l3.518-3.568 3.554 3.521 1.431-1.43-3.566-3.523 3.535-3.568-1.431-1.432-3.539 3.583-3.581-3.457-1.418 1.418 3.585 3.473-3.507 3.566 1.419 1.417z"/>',warning:'<path d="M12 0l-12 12 12 12 12-12-12-12zm-1 6h2v8h-2v-8zm1 12.25c-.69 0-1.25-.56-1.25-1.25s.56-1.25 1.25-1.25 1.25.56 1.25 1.25-.56 1.25-1.25 1.25z"/>',info:'<path d="M12 2c5.514 0 10 4.486 10 10s-4.486 10-10 10-10-4.486-10-10 4.486-10 10-10zm0-2c-6.627 0-12 5.373-12 12s5.373 12 12 12 12-5.373 12-12-5.373-12-12-12zm-2.033 16.01c.564-1.789 1.632-3.932 1.821-4.474.273-.787-.211-1.136-1.74.209l-.34-.64c1.744-1.897 5.335-2.326 4.113.613-.763 1.835-1.309 3.074-1.621 4.03-.455 1.393.694.828 1.819-.211.153.25.203.331.356.619-2.498 2.378-5.271 2.588-4.408-.146zm4.742-8.169c-.532.453-1.32.443-1.761-.022-.441-.465-.367-1.208.164-1.661.532-.453 1.32-.442 1.761.022.439.466.367 1.209-.164 1.661z"/>',bin:'<path d="M9 19c0 .552-.448 1-1 1s-1-.448-1-1v-10c0-.552.448-1 1-1s1 .448 1 1v10zm4 0c0 .552-.448 1-1 1s-1-.448-1-1v-10c0-.552.448-1 1-1s1 .448 1 1v10zm4 0c0 .552-.448 1-1 1s-1-.448-1-1v-10c0-.552.448-1 1-1s1 .448 1 1v10zm5-17v2h-20v-2h5.711c.9 0 1.631-1.099 1.631-2h5.315c0 .901.73 2 1.631 2h5.712zm-3 4v16h-14v-16h-2v18h18v-18h-2z"/>',save:'<path d="M13 3h2.996v5h-2.996v-5zm11 1v20h-24v-24h20l4 4zm-17 5h10v-7h-10v7zm15-4.171l-2.828-2.829h-.172v9h-14v-9h-3v20h20v-17.171zm-3 10.171h-14v1h14v-1zm0 2h-14v1h14v-1zm0 2h-14v1h14v-1z"/>',cookie:'<path d="M23.999 12.149c-.049 3.834-1.893 7.223-4.706 9.378-1.993 1.53-4.485 2.449-7.198 2.473-6.464.057-12.051-5.107-12.095-12 3.966 1.066 7.682-1.993 6-6 4.668.655 6.859-2.389 6.077-6 6.724.064 11.999 5.542 11.922 12.149zm-15.576-4.123c-.065 3.393-2.801 5.868-6.182 6.166 1.008 4.489 5.015 7.807 9.759 7.807 5.262 0 9.576-4.072 9.97-9.229.369-4.818-2.755-9.357-7.796-10.534-.277 2.908-2.381 5.357-5.751 5.79zm5.077 8.974c.828 0 1.5.672 1.5 1.5s-.672 1.5-1.5 1.5-1.5-.672-1.5-1.5.672-1.5 1.5-1.5zm-5.5-2.853c1.104 0 2 .896 2 2s-.896 2-2 2-2-.896-2-2 .896-2 2-2zm10-2.147c1.104 0 2 .896 2 2s-.896 2-2 2-2-.896-2-2 .896-2 2-2zm-5 0c.552 0 1 .448 1 1s-.448 1-1 1-1-.448-1-1 .448-1 1-1zm2.5-5c.828 0 1.5.672 1.5 1.5s-.672 1.5-1.5 1.5-1.5-.672-1.5-1.5.672-1.5 1.5-1.5zm-12.5 0c.552 0 1 .448 1 1s-.448 1-1 1-1-.448-1-1 .448-1 1-1zm-1.5-4c.828 0 1.5.672 1.5 1.5s-.672 1.5-1.5 1.5-1.5-.672-1.5-1.5.672-1.5 1.5-1.5zm6-2c.828 0 1.5.672 1.5 1.5s-.672 1.5-1.5 1.5-1.5-.672-1.5-1.5.672-1.5 1.5-1.5zm-3.5-1c.552 0 1 .448 1 1s-.448 1-1 1-1-.448-1-1 .448-1 1-1z"/>'},h={add:function(e,t){return g("add",e,t)},check:function(e,t){return g("check",e,t)},success:function(e,t){return g("success",e,t)},cross:function(e,t){return g("cross",e,t)},crossArrow:function(e,t){return g("crossArrow",e,t)},error:function(e,t){return g("error",e,t)},warning:function(e,t){return g("warning",e,t)},info:function(e,t){return g("info",e,t)},bin:function(e,t){return g("bin",e,t)},save:function(e,t){return g("save",e,t)},cookie:function(e,t){return g("cookie",e,t)}},g=function(e,t,n){switch(t=t||24,e){case"add":case"check":case"success":n=n||"#149414";break;case"cross":case"crossArrow":case"error":case"bin":n=n||"#CC0000";break;case"warning":n=n||"#FFA500";break;case"info":case"save":n=n||"#195e9f";break;case"cookie":n=n||"#984e0a";break;default:return'<b style="color:darkred">Error:</b> <b>Icon not found for <em>'+e+"</em></b>"}return'<svg xmlns="http://www.w3.org/2000/svg" class="ccfw-icon ccfw-icon__'+e+'" width="'+t+'" height="'+t+'" fill="'+n+'" viewBox="0 0 24 24">'+m[e]+" </svg>"},b=function(e,t){return jQuery("<"+e+"/>",t||null)},y=function(e){return b("span",{class:"ccfw-group__heading"}).html("Group name: <span>"+(e||"")+"</span>")},w=function(e){var t=e||"-- Create an item --";return b("option",{value:(e=e||"").toLowerCase()}).text(t)},x=function(e,t){return b("input",{type:"text",name:e,placeholder:t||"",class:"ccfw-cookie-row__"+e})},k=function(e,t,n,r){var o=b("button",{type:"button",class:e});return n&&o.append(n),r&&o.attr("title",r),t&&o.append(t),o};function O(e){return A.manage.allowlistIdsValid()?A.sectionsChanged&&(removeEventListener("beforeunload",A.manage.beforeUnload,{capture:!0}),jQuery("html, body").animate({scrollTop:30},"fast"),!A.sectionsStored)?(e.preventDefault(),_.form.post(A.sections),!1):void 0:(e.preventDefault(),!1)}function j(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=null==e?null:"undefined"!=typeof Symbol&&e[Symbol.iterator]||e["@@iterator"];if(null==n)return;var r,o,i=[],a=!0,c=!1;try{for(n=n.call(e);!(a=(r=n.next()).done)&&(i.push(r.value),!t||i.length!==t);a=!0);}catch(e){c=!0,o=e}finally{try{a||null==n.return||n.return()}finally{if(c)throw o}}return i}(e,t)||function(e,t){if(!e)return;if("string"==typeof e)return E(e,t);var n=Object.prototype.toString.call(e).slice(8,-1);"Object"===n&&e.constructor&&(n=e.constructor.name);if("Map"===n||"Set"===n)return Array.from(e);if("Arguments"===n||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n))return E(e,t)}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}()}function E(e,t){(null==t||t>e.length)&&(t=e.length);for(var n=0,r=new Array(t);n<t;n++)r[n]=e[n];return r}var A={appContainer:"ccfw-cookie-management",sectionId:"ccfw-section-select-container",sections:{},sectionsStored:!1,sectionsLoaded:!1,sectionsChanged:!1,building:!1,cookieRow:{name:"",description:"",expiry:""},debug:{container:"debugging",preContainer:"debug-settings",import:{container:"ccfw-import-container",textarea:"ccfw-import-cookie-object"},active:!1,checkbox:"cookie-management-debug",init:function(){jQuery.migrateMute=!0,jQuery("#"+A.debug.checkbox).is(":checked")&&(A.debug.active=!0,A.debug.make())},make:function(){var e,t,n=b("div",{id:A.debug.container}),r=b("div",{class:A.debug.container+"-icons"});r.append((t="<h3>Available icons</h3>",e=(e=8)||8,Object.entries(h).forEach((function(n,r){t+=n[1]()+((r+1)%e==0?"<br><br>":" &nbsp; ")})),t)),n.append(r),n.append('<div id ="'+A.debug.preContainer+'"/>'),jQuery("#"+A.appContainer).before(n),A.debug.output()},output:function(){if(A.debug.active){var e=JSON.stringify(A.sections,void 0,4);if(e.length<5){var t=b("div",{id:A.debug.import.container});t.append("<h3>Import cookies</h3>"),t.append(b("textarea",{class:A.debug.import.textarea})),jQuery("#"+A.debug.preContainer).html(t),v.import()}else{document.getElementById(A.debug.preContainer).innerHTML="<pre> "+e+" </pre>",v.copy()}}}},create:{app:function(){var e=jQuery("#"+A.appContainer);e.html(A.manage.sectionSelect().append(A.manage.appInfoButton())),e.prepend('<div class="ccfw-ajax"> <div class="loader"> </div> </div>')},section:function(e){var t=A.building?"#"+A.sectionId+"-select":this,n=jQuery(t),r=A.building?e:n.val(),o=b("div",{id:"ccfw-section__"+r,class:"ccfw-section"}).data("id",r),i=b("button",{type:"button",id:"ccfw-"+r+"-section-remove",class:"ccfw-section-remove",title:"Click to remove the "+r+" section and all related cookies"});_.section.add(r),i.append(h.bin("20")),o.append(i),o.append(b("h2").text(S(r)+" Cookies"));var a=b("div",{class:"ccfw-group"}),c=b("div",{class:"ccfw-group-define"}),s=b("button",{type:"button",class:"ccfw-"+r+"-group-save",title:"Create a group"});s.append(h.add(22)),a.append(y()),c.append(x("ccfw-group__input")),c.append(s),a.append(c),o.append(a),jQuery("#"+A.appContainer).append(o),A.building||a.find('input[name="ccfw-group__input"]').focus(),f.save(r,A.create.group),jQuery("#ccfw-"+r+"-section-remove").on("click",A.manage.deleteSection),n.find("option[value="+r+"]").remove(),A.manage.hideShowSectionSelect()},group:function(e){var t=A.building?".ccfw-"+e.section+"-group-save":this,n=jQuery(t).closest("div.ccfw-section"),o=jQuery(t).closest("div.ccfw-group"),i=A.building?e.group:o.find("input").val(),a=D(i),c=n.data("id"),s="",u="marketing"===c?x("gtm-allowlist-id","GTM Allowlist ID"):null,p=x("ccfw-group-description","Enter "+i+" group blurb");if(i)if(A.building||!_.group.exists(c,a)){_.group.add(c,i),o.data("id",a),o.addClass("ccfw-group__"+a),o.html(y(i+" "+h.success(18))),o.append(k("ccfw-group-remove",h.bin(16,"grey"))),u&&o.append(u),(s=b("div",{class:"ccfw-group__description"})).append(p),s.append(k("ccfw-cookie-row-add ccfw-cookie-row-add-"+a,h.add(20),"Add cookie")),o.append(s);var l=b("div",{class:"ccfw-cookie-container"});if(!A.building){var v=A.create.cookieRow(0);l.append(v),_.cookies.add(c,i,"row_0"),v.find('input[name="name"]').focus()}o.append(l),n.append(A.manage.group(c)),f.save(c,A.create.group),f.remove(A.manage.removeGroup),f.allowlist.save(A.save.allowListId),f.description.save(A.save.description),d.add(A.create.cookie),d.remove(A.manage.removeRow),d.save(A.save.cookieData),A.building&&(u&&u.val(e.id),p.val(e.description))}else r.alert('The group "'+i+'" already exists. Please chose another name.');else console.warn("Group name was empty.")},cookie:function(e,t){var n=A.building?"#ccfw-section__"+t.section+" .ccfw-cookie-row-add-"+D(t.group):this,r=jQuery(n).parent().siblings(".ccfw-cookie-container"),o=A.building?e:r.find(".ccfw-cookie-row").length,i=A.manage.locationData(jQuery(n)),a=!1;if(!A.building)for(;!1===a;){if(!_.cookies.row.exists(i.section,i.group,"row_"+o)){a=!0;break}o++}var c=A.create.cookieRow(o);r.append(c),A.building?(c.find('input[name="name"]').val(t.row.name),c.find('input[name="description"]').val(t.row.description),c.find('input[name="expiry"]').val(t.row.expiry)):c.find('input[name="name"]').focus(),_.cookies.add(i.section,i.group,"row_"+o),d.save(A.save.cookieData),d.remove(A.manage.removeRow)},cookieRow:function(e){var t=b("div",{class:"ccfw-cookie-row"}).data("id","row_"+e);return t.append(x("name","Name")),t.append(x("description","Description")),t.append(x("expiry","Expiry or URL")),t.append(k("ccfw-cookie-row-remove",h.crossArrow(30))),t}},save:{allowListId:function(){var e=jQuery(this),t=A.manage.locationData(e);e.removeClass("ccfw-form-input-invalid"),e.next("svg.ccfw-icon__warning").remove(),_.group.allowlistId.save(t,e.val())},description:function(){var e=jQuery(this),t=A.manage.locationData(e);_.group.description.save(t,e.val())},cookieData:function(){var e=jQuery(this),t=A.manage.locationData(e),n=e.parent().data("id"),r=e.prop("name");_.cookies.update(t.section,t.group,n,r,e.val())}},manage:{hideShowSectionSelect:function(){jQuery("#"+A.sectionId+" option").length>1?jQuery("#"+A.sectionId).show():jQuery("#"+A.sectionId).hide()},deleteSection:function(){var e=jQuery(this).parent("div").data("id");r.confirm("\nAre you sure you want to remove "+e+" cookies?")&&(jQuery("#ccfw-section__"+e).remove(),_.section.remove(e),jQuery("#"+A.sectionId+" select").append(w(S(e))),A.manage.hideShowSectionSelect())},inputFocus:function(e){var t=jQuery(this).closest(".ccfw-group");"focusin"===e.type?t.removeClass("ccfw-group__active-opacity"):t.addClass("ccfw-group__active-opacity")},locationData:function(e){return{section:e.closest("div.ccfw-section").data("id"),group:e.closest("div.ccfw-group").data("id")}},appInfoButton:function(){return k("app-info",h.info(),null,"Choose an option from the dropdown to create a cookie section. The section will allow you to define groups of cookies.")},removeGroup:function(){var e=A.manage.locationData(jQuery(this));r.confirm("\nAre you sure you want to remove the "+e.group+" group from "+e.section+" cookies?")&&(jQuery(".ccfw-group__"+e.group).remove(),_.group.remove(e.section,e.group))},removeRow:function(){var e=A.manage.locationData(jQuery(this)),t=jQuery(this).parent(),n=t.data("id");_.cookies.remove(e.section,e.group,n),t.remove()},group:function(e){var t=b("div",{class:"ccfw-group"}),n=b("div",{class:"ccfw-group-define"}),r=b("button",{type:"button",class:"ccfw-"+e+"-group-save",title:"Create a group"});return r.append(h.add(22)),t.append(y()),n.append(x("ccfw-group__input")),n.append(r),t.append(n),t},sectionSelect:function(e){var t=["Third-party","Essential","Marketing"];return e&&e.forEach((function(e){t.push(e)})),function(e,t,n){var r=b("select",{name:e,id:e+"-select"}),o=b("div",{id:e});return r.append(w("Create a section")),t.forEach((function(e){r.append(w(e))})),r.on("change",n),o.html(r),o}(A.sectionId,t.reverse(),A.create.section)},beforeUnload:function(e){return e.preventDefault(),e.returnValue="Changes have been made to cookie content. Are you sure you want to leave?"},allowlistIdsValid:function(){for(var e=!0,t=0,n=Object.entries(A.sections);t<n.length;t++){var o=j(n[t],2),i=o[0],a=o[1];if("marketing"===i)for(var c=0,s=Object.entries(a);c<s.length;c++){var u=j(s[c],2),p=u[0],f=u[1];if(!T(i,p)){var d=jQuery("#ccfw-section__"+i+" .ccfw-group__"+p+" .ccfw-cookie-row__gtm-allowlist-id");d.after(h.warning(18)),d.addClass("ccfw-form-input-invalid"),r.alert("The "+f.name+" group has an invalid GTM Allowlist ID."),e=!1}}}return e}}},C=function(){A.sectionsChanged=!0,addEventListener("beforeunload",A.manage.beforeUnload,{capture:!0})},T=function(e,t){var n=A.sections[e][t];return!!n.hasOwnProperty("allowlistID")&&0!==n.allowlistID.length},S=function(e){return"string"!=typeof e?"":e.charAt(0).toUpperCase()+e.slice(1)},D=function(e){return e.toLowerCase().replace(/ /g,"-").replace(/[-]+/g,"-").replace(/[^\w-]+/g,"")},_={section:{add:function(e){if(A.building)return!1;A.sections[e]={},C(),A.debug.output()},remove:function(e){if(A.building)return!1;delete A.sections[e],C(),A.debug.output()}},group:{add:function(e,t){if(A.building)return!1;C();var n=D(t);A.sections[e][n]={name:t},A.sections[e][n].cookies={row_0:{name:"",description:"",expiry:""}},A.debug.output()},remove:function(e,t){if(A.building)return!1;C(),delete A.sections[e][t],A.debug.output()},exists:function(e,t){return A.sections[e].hasOwnProperty(t)},allowlistId:{save:function(e,t){if(A.building)return!1;C(),A.sections[e.section][D(e.group)].allowlistID=t.trim(),A.debug.output()}},description:{save:function(e,t){if(A.building)return!1;C(),A.sections[e.section][D(e.group)].description=t.trim(),A.debug.output()}}},cookies:{add:function(e,t,n){if(A.building)return!1;C(),A.sections[e][D(t)].cookies[n]={name:"",description:"",expiry:""},A.debug.output()},update:function(e,t,n,r,o){if(A.building)return!1;C(),A.sections[e][D(t)].cookies[n][r]=o.trim(),A.debug.output()},remove:function(e,t,n){if(A.building)return!1;C(),delete A.sections[e][D(t)].cookies[n],A.debug.output()},row:{exists:function(e,t,n){return t=D(t),A.sections[e][t].cookies||(A.sections[e][t].cookies={}),A.sections[e][t].cookies.hasOwnProperty(n)}},get:function(){return i()}},form:{post:function(e){return o(e)}}},L=n(861),M=n(379),I=n.n(M),Q=n(535),z={insert:"head",singleton:!1};I()(Q.Z,z);Q.Z.locals;jQuery((function(){s.init(),A.debug.init(),l.input.focusBlur(A.manage.inputFocus),v.toggle(),l.submit(O),A.create.app(),jQuery("h2:contains(Cookie Management Settings)").css({marginBottom:"1.68em"}).prepend(h.cookie(28)),(0,L.ZP)(".app-info",{aria:{content:"describedby"},content:function(e){return e.getAttribute("title")},offset:[-18,10]})}))},502:()=>{function e(t){return(e="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(t)}Object.keys||(Object.keys=function(){"use strict";var t=Object.prototype.hasOwnProperty,n=!{toString:null}.propertyIsEnumerable("toString"),r=["toString","toLocaleString","valueOf","hasOwnProperty","isPrototypeOf","propertyIsEnumerable","constructor"],o=r.length;return function(i){if("function"!=typeof i&&("object"!==e(i)||null===i))throw new TypeError("Object.keys called on non-object");var a,c,s=[];for(a in i)t.call(i,a)&&s.push(a);if(n)for(c=0;c<o;c++)t.call(i,r[c])&&s.push(r[c]);return s}}())},108:()=>{jQuery((function(e){function t(e,t){var n=new URLSearchParams(window.location.search);return!t&&n.has(e)?n.get(e):!!e&&(n.set(e,t),window.history&&window.history.replaceState({},"","".concat(location.pathname,"?").concat(n)),window.location.pathname+window.location.search)}function n(t){return t.preventDefault(),r(e(this).attr("href")),!1}function r(n){var r,o,i=e(".nav-tab-wrapper a");return(n=n?e(".nav-tab-wrapper a[href='"+n+"']"):i.eq(0)).attr("href")||(n=i.eq(0)),r=n.attr("href").split("#")[1],n.parent().find("a").removeClass("nav-tab-active"),n.addClass("nav-tab-active"),e(".ccfw-component-settings-section").hide(),e("div#"+r).fadeIn(),o=t("ccfw-tab",r),e('input[name="_wp_http_referer"]').val(o),!1}if(e(".settings_page_cookie-compliance-for-wordpress-settings").length>0){e(".nav-tab-wrapper").on("click","a",n),e(".ccfw-inline-tab-link").on("click",n);var o=t("ccfw-tab");r(o?"#"+o:null)}})),Object.entries||(Object.entries=function(e){for(var t=Object.keys(e),n=t.length,r=new Array(n);n--;)r[n]=[t[n],e[t[n]]];return r})},535:(e,t,n)=>{"use strict";n.d(t,{Z:()=>i});var r=n(645),o=n.n(r)()((function(e){return e[1]}));o.push([e.id,'.tippy-box[data-animation=fade][data-state=hidden]{opacity:0}[data-tippy-root]{max-width:calc(100vw - 10px)}.tippy-box{background-color:#333;border-radius:4px;color:#fff;font-size:14px;line-height:1.4;outline:0;position:relative;transition-property:transform,visibility,opacity}.tippy-box[data-placement^=top]>.tippy-arrow{bottom:0}.tippy-box[data-placement^=top]>.tippy-arrow:before{border-top-color:initial;border-width:8px 8px 0;bottom:-7px;left:0;transform-origin:center top}.tippy-box[data-placement^=bottom]>.tippy-arrow{top:0}.tippy-box[data-placement^=bottom]>.tippy-arrow:before{border-bottom-color:initial;border-width:0 8px 8px;left:0;top:-7px;transform-origin:center bottom}.tippy-box[data-placement^=left]>.tippy-arrow{right:0}.tippy-box[data-placement^=left]>.tippy-arrow:before{border-left-color:initial;border-width:8px 0 8px 8px;right:-7px;transform-origin:center left}.tippy-box[data-placement^=right]>.tippy-arrow{left:0}.tippy-box[data-placement^=right]>.tippy-arrow:before{border-right-color:initial;border-width:8px 8px 8px 0;left:-7px;transform-origin:center right}.tippy-box[data-inertia][data-state=visible]{transition-timing-function:cubic-bezier(.54,1.5,.38,1.11)}.tippy-arrow{color:#333;height:16px;width:16px}.tippy-arrow:before{border-color:transparent;border-style:solid;content:"";position:absolute}.tippy-content{padding:5px 9px;position:relative;z-index:1}',""]);const i=o},645:e=>{"use strict";e.exports=function(e){var t=[];return t.toString=function(){return this.map((function(t){var n=e(t);return t[2]?"@media ".concat(t[2]," {").concat(n,"}"):n})).join("")},t.i=function(e,n,r){"string"==typeof e&&(e=[[null,e,""]]);var o={};if(r)for(var i=0;i<this.length;i++){var a=this[i][0];null!=a&&(o[a]=!0)}for(var c=0;c<e.length;c++){var s=[].concat(e[c]);r&&o[s[0]]||(n&&(s[2]?s[2]="".concat(n," and ").concat(s[2]):s[2]=n),t.push(s))}},t}},379:(e,t,n)=>{"use strict";var r,o=function(){return void 0===r&&(r=Boolean(window&&document&&document.all&&!window.atob)),r},i=function(){var e={};return function(t){if(void 0===e[t]){var n=document.querySelector(t);if(window.HTMLIFrameElement&&n instanceof window.HTMLIFrameElement)try{n=n.contentDocument.head}catch(e){n=null}e[t]=n}return e[t]}}(),a=[];function c(e){for(var t=-1,n=0;n<a.length;n++)if(a[n].identifier===e){t=n;break}return t}function s(e,t){for(var n={},r=[],o=0;o<e.length;o++){var i=e[o],s=t.base?i[0]+t.base:i[0],u=n[s]||0,p="".concat(s," ").concat(u);n[s]=u+1;var f=c(p),d={css:i[1],media:i[2],sourceMap:i[3]};-1!==f?(a[f].references++,a[f].updater(d)):a.push({identifier:p,updater:h(d,t),references:1}),r.push(p)}return r}function u(e){var t=document.createElement("style"),r=e.attributes||{};if(void 0===r.nonce){var o=n.nc;o&&(r.nonce=o)}if(Object.keys(r).forEach((function(e){t.setAttribute(e,r[e])})),"function"==typeof e.insert)e.insert(t);else{var a=i(e.insert||"head");if(!a)throw new Error("Couldn't find a style target. This probably means that the value for the 'insert' parameter is invalid.");a.appendChild(t)}return t}var p,f=(p=[],function(e,t){return p[e]=t,p.filter(Boolean).join("\n")});function d(e,t,n,r){var o=n?"":r.media?"@media ".concat(r.media," {").concat(r.css,"}"):r.css;if(e.styleSheet)e.styleSheet.cssText=f(t,o);else{var i=document.createTextNode(o),a=e.childNodes;a[t]&&e.removeChild(a[t]),a.length?e.insertBefore(i,a[t]):e.appendChild(i)}}function l(e,t,n){var r=n.css,o=n.media,i=n.sourceMap;if(o?e.setAttribute("media",o):e.removeAttribute("media"),i&&"undefined"!=typeof btoa&&(r+="\n/*# sourceMappingURL=data:application/json;base64,".concat(btoa(unescape(encodeURIComponent(JSON.stringify(i))))," */")),e.styleSheet)e.styleSheet.cssText=r;else{for(;e.firstChild;)e.removeChild(e.firstChild);e.appendChild(document.createTextNode(r))}}var v=null,m=0;function h(e,t){var n,r,o;if(t.singleton){var i=m++;n=v||(v=u(t)),r=d.bind(null,n,i,!1),o=d.bind(null,n,i,!0)}else n=u(t),r=l.bind(null,n,t),o=function(){!function(e){if(null===e.parentNode)return!1;e.parentNode.removeChild(e)}(n)};return r(e),function(t){if(t){if(t.css===e.css&&t.media===e.media&&t.sourceMap===e.sourceMap)return;r(e=t)}else o()}}e.exports=function(e,t){(t=t||{}).singleton||"boolean"==typeof t.singleton||(t.singleton=o());var n=s(e=e||[],t);return function(e){if(e=e||[],"[object Array]"===Object.prototype.toString.call(e)){for(var r=0;r<n.length;r++){var o=c(n[r]);a[o].references--}for(var i=s(e,t),u=0;u<n.length;u++){var p=c(n[u]);0===a[p].references&&(a[p].updater(),a.splice(p,1))}n=i}}}},861:(e,t,n)=>{"use strict";function r(e){var t=e.getBoundingClientRect();return{width:t.width,height:t.height,top:t.top,right:t.right,bottom:t.bottom,left:t.left,x:t.left,y:t.top}}function o(e){if(null==e)return window;if("[object Window]"!==e.toString()){var t=e.ownerDocument;return t&&t.defaultView||window}return e}function i(e){var t=o(e);return{scrollLeft:t.pageXOffset,scrollTop:t.pageYOffset}}function a(e){return e instanceof o(e).Element||e instanceof Element}function c(e){return e instanceof o(e).HTMLElement||e instanceof HTMLElement}function s(e){return"undefined"!=typeof ShadowRoot&&(e instanceof o(e).ShadowRoot||e instanceof ShadowRoot)}function u(e){return e?(e.nodeName||"").toLowerCase():null}function p(e){return((a(e)?e.ownerDocument:e.document)||window.document).documentElement}function f(e){return r(p(e)).left+i(e).scrollLeft}function d(e){return o(e).getComputedStyle(e)}function l(e){var t=d(e),n=t.overflow,r=t.overflowX,o=t.overflowY;return/auto|scroll|overlay|hidden/.test(n+o+r)}function v(e,t,n){void 0===n&&(n=!1);var a,s,d=p(t),v=r(e),m=c(t),h={scrollLeft:0,scrollTop:0},g={x:0,y:0};return(m||!m&&!n)&&(("body"!==u(t)||l(d))&&(h=(a=t)!==o(a)&&c(a)?{scrollLeft:(s=a).scrollLeft,scrollTop:s.scrollTop}:i(a)),c(t)?((g=r(t)).x+=t.clientLeft,g.y+=t.clientTop):d&&(g.x=f(d))),{x:v.left+h.scrollLeft-g.x,y:v.top+h.scrollTop-g.y,width:v.width,height:v.height}}function m(e){var t=r(e),n=e.offsetWidth,o=e.offsetHeight;return Math.abs(t.width-n)<=1&&(n=t.width),Math.abs(t.height-o)<=1&&(o=t.height),{x:e.offsetLeft,y:e.offsetTop,width:n,height:o}}function h(e){return"html"===u(e)?e:e.assignedSlot||e.parentNode||(s(e)?e.host:null)||p(e)}function g(e){return["html","body","#document"].indexOf(u(e))>=0?e.ownerDocument.body:c(e)&&l(e)?e:g(h(e))}function b(e,t){var n;void 0===t&&(t=[]);var r=g(e),i=r===(null==(n=e.ownerDocument)?void 0:n.body),a=o(r),c=i?[a].concat(a.visualViewport||[],l(r)?r:[]):r,s=t.concat(c);return i?s:s.concat(b(h(c)))}function y(e){return["table","td","th"].indexOf(u(e))>=0}function w(e){return c(e)&&"fixed"!==d(e).position?e.offsetParent:null}function x(e){for(var t=o(e),n=w(e);n&&y(n)&&"static"===d(n).position;)n=w(n);return n&&("html"===u(n)||"body"===u(n)&&"static"===d(n).position)?t:n||function(e){var t=-1!==navigator.userAgent.toLowerCase().indexOf("firefox");if(-1!==navigator.userAgent.indexOf("Trident")&&c(e)&&"fixed"===d(e).position)return null;for(var n=h(e);c(n)&&["html","body"].indexOf(u(n))<0;){var r=d(n);if("none"!==r.transform||"none"!==r.perspective||"paint"===r.contain||-1!==["transform","perspective"].indexOf(r.willChange)||t&&"filter"===r.willChange||t&&r.filter&&"none"!==r.filter)return n;n=n.parentNode}return null}(e)||t}n.d(t,{ZP:()=>rt});var k="top",O="bottom",j="right",E="left",A="auto",C=[k,O,j,E],T="start",S="end",D="viewport",_="popper",L=C.reduce((function(e,t){return e.concat([t+"-"+T,t+"-"+S])}),[]),M=[].concat(C,[A]).reduce((function(e,t){return e.concat([t,t+"-"+T,t+"-"+S])}),[]),I=["beforeRead","read","afterRead","beforeMain","main","afterMain","beforeWrite","write","afterWrite"];function Q(e){var t=new Map,n=new Set,r=[];function o(e){n.add(e.name),[].concat(e.requires||[],e.requiresIfExists||[]).forEach((function(e){if(!n.has(e)){var r=t.get(e);r&&o(r)}})),r.push(e)}return e.forEach((function(e){t.set(e.name,e)})),e.forEach((function(e){n.has(e.name)||o(e)})),r}var z={placement:"bottom",modifiers:[],strategy:"absolute"};function P(){for(var e=arguments.length,t=new Array(e),n=0;n<e;n++)t[n]=arguments[n];return!t.some((function(e){return!(e&&"function"==typeof e.getBoundingClientRect)}))}function B(e){void 0===e&&(e={});var t=e,n=t.defaultModifiers,r=void 0===n?[]:n,o=t.defaultOptions,i=void 0===o?z:o;return function(e,t,n){void 0===n&&(n=i);var o,c,s={placement:"bottom",orderedModifiers:[],options:Object.assign({},z,i),modifiersData:{},elements:{reference:e,popper:t},attributes:{},styles:{}},u=[],p=!1,f={state:s,setOptions:function(n){d(),s.options=Object.assign({},i,s.options,n),s.scrollParents={reference:a(e)?b(e):e.contextElement?b(e.contextElement):[],popper:b(t)};var o=function(e){var t=Q(e);return I.reduce((function(e,n){return e.concat(t.filter((function(e){return e.phase===n})))}),[])}(function(e){var t=e.reduce((function(e,t){var n=e[t.name];return e[t.name]=n?Object.assign({},n,t,{options:Object.assign({},n.options,t.options),data:Object.assign({},n.data,t.data)}):t,e}),{});return Object.keys(t).map((function(e){return t[e]}))}([].concat(r,s.options.modifiers)));return s.orderedModifiers=o.filter((function(e){return e.enabled})),s.orderedModifiers.forEach((function(e){var t=e.name,n=e.options,r=void 0===n?{}:n,o=e.effect;if("function"==typeof o){var i=o({state:s,name:t,instance:f,options:r}),a=function(){};u.push(i||a)}})),f.update()},forceUpdate:function(){if(!p){var e=s.elements,t=e.reference,n=e.popper;if(P(t,n)){s.rects={reference:v(t,x(n),"fixed"===s.options.strategy),popper:m(n)},s.reset=!1,s.placement=s.options.placement,s.orderedModifiers.forEach((function(e){return s.modifiersData[e.name]=Object.assign({},e.data)}));for(var r=0;r<s.orderedModifiers.length;r++)if(!0!==s.reset){var o=s.orderedModifiers[r],i=o.fn,a=o.options,c=void 0===a?{}:a,u=o.name;"function"==typeof i&&(s=i({state:s,options:c,name:u,instance:f})||s)}else s.reset=!1,r=-1}}},update:(o=function(){return new Promise((function(e){f.forceUpdate(),e(s)}))},function(){return c||(c=new Promise((function(e){Promise.resolve().then((function(){c=void 0,e(o())}))}))),c}),destroy:function(){d(),p=!0}};if(!P(e,t))return f;function d(){u.forEach((function(e){return e()})),u=[]}return f.setOptions(n).then((function(e){!p&&n.onFirstUpdate&&n.onFirstUpdate(e)})),f}}var R={passive:!0};const V={name:"eventListeners",enabled:!0,phase:"write",fn:function(){},effect:function(e){var t=e.state,n=e.instance,r=e.options,i=r.scroll,a=void 0===i||i,c=r.resize,s=void 0===c||c,u=o(t.elements.popper),p=[].concat(t.scrollParents.reference,t.scrollParents.popper);return a&&p.forEach((function(e){e.addEventListener("scroll",n.update,R)})),s&&u.addEventListener("resize",n.update,R),function(){a&&p.forEach((function(e){e.removeEventListener("scroll",n.update,R)})),s&&u.removeEventListener("resize",n.update,R)}},data:{}};function N(e){return e.split("-")[0]}function H(e){return e.split("-")[1]}function U(e){return["top","bottom"].indexOf(e)>=0?"x":"y"}function W(e){var t,n=e.reference,r=e.element,o=e.placement,i=o?N(o):null,a=o?H(o):null,c=n.x+n.width/2-r.width/2,s=n.y+n.height/2-r.height/2;switch(i){case k:t={x:c,y:n.y-r.height};break;case O:t={x:c,y:n.y+n.height};break;case j:t={x:n.x+n.width,y:s};break;case E:t={x:n.x-r.width,y:s};break;default:t={x:n.x,y:n.y}}var u=i?U(i):null;if(null!=u){var p="y"===u?"height":"width";switch(a){case T:t[u]=t[u]-(n[p]/2-r[p]/2);break;case S:t[u]=t[u]+(n[p]/2-r[p]/2)}}return t}const q={name:"popperOffsets",enabled:!0,phase:"read",fn:function(e){var t=e.state,n=e.name;t.modifiersData[n]=W({reference:t.rects.reference,element:t.rects.popper,strategy:"absolute",placement:t.placement})},data:{}};var F=Math.max,K=Math.min,$=Math.round,J={top:"auto",right:"auto",bottom:"auto",left:"auto"};function G(e){var t,n=e.popper,r=e.popperRect,i=e.placement,a=e.offsets,c=e.position,s=e.gpuAcceleration,u=e.adaptive,f=e.roundOffsets,l=!0===f?function(e){var t=e.x,n=e.y,r=window.devicePixelRatio||1;return{x:$($(t*r)/r)||0,y:$($(n*r)/r)||0}}(a):"function"==typeof f?f(a):a,v=l.x,m=void 0===v?0:v,h=l.y,g=void 0===h?0:h,b=a.hasOwnProperty("x"),y=a.hasOwnProperty("y"),w=E,A=k,C=window;if(u){var T=x(n),S="clientHeight",D="clientWidth";T===o(n)&&"static"!==d(T=p(n)).position&&(S="scrollHeight",D="scrollWidth"),T=T,i===k&&(A=O,g-=T[S]-r.height,g*=s?1:-1),i===E&&(w=j,m-=T[D]-r.width,m*=s?1:-1)}var _,L=Object.assign({position:c},u&&J);return s?Object.assign({},L,((_={})[A]=y?"0":"",_[w]=b?"0":"",_.transform=(C.devicePixelRatio||1)<2?"translate("+m+"px, "+g+"px)":"translate3d("+m+"px, "+g+"px, 0)",_)):Object.assign({},L,((t={})[A]=y?g+"px":"",t[w]=b?m+"px":"",t.transform="",t))}const Z={name:"applyStyles",enabled:!0,phase:"write",fn:function(e){var t=e.state;Object.keys(t.elements).forEach((function(e){var n=t.styles[e]||{},r=t.attributes[e]||{},o=t.elements[e];c(o)&&u(o)&&(Object.assign(o.style,n),Object.keys(r).forEach((function(e){var t=r[e];!1===t?o.removeAttribute(e):o.setAttribute(e,!0===t?"":t)})))}))},effect:function(e){var t=e.state,n={popper:{position:t.options.strategy,left:"0",top:"0",margin:"0"},arrow:{position:"absolute"},reference:{}};return Object.assign(t.elements.popper.style,n.popper),t.styles=n,t.elements.arrow&&Object.assign(t.elements.arrow.style,n.arrow),function(){Object.keys(t.elements).forEach((function(e){var r=t.elements[e],o=t.attributes[e]||{},i=Object.keys(t.styles.hasOwnProperty(e)?t.styles[e]:n[e]).reduce((function(e,t){return e[t]="",e}),{});c(r)&&u(r)&&(Object.assign(r.style,i),Object.keys(o).forEach((function(e){r.removeAttribute(e)})))}))}},requires:["computeStyles"]};var X={left:"right",right:"left",bottom:"top",top:"bottom"};function Y(e){return e.replace(/left|right|bottom|top/g,(function(e){return X[e]}))}var ee={start:"end",end:"start"};function te(e){return e.replace(/start|end/g,(function(e){return ee[e]}))}function ne(e,t){var n=t.getRootNode&&t.getRootNode();if(e.contains(t))return!0;if(n&&s(n)){var r=t;do{if(r&&e.isSameNode(r))return!0;r=r.parentNode||r.host}while(r)}return!1}function re(e){return Object.assign({},e,{left:e.x,top:e.y,right:e.x+e.width,bottom:e.y+e.height})}function oe(e,t){return t===D?re(function(e){var t=o(e),n=p(e),r=t.visualViewport,i=n.clientWidth,a=n.clientHeight,c=0,s=0;return r&&(i=r.width,a=r.height,/^((?!chrome|android).)*safari/i.test(navigator.userAgent)||(c=r.offsetLeft,s=r.offsetTop)),{width:i,height:a,x:c+f(e),y:s}}(e)):c(t)?function(e){var t=r(e);return t.top=t.top+e.clientTop,t.left=t.left+e.clientLeft,t.bottom=t.top+e.clientHeight,t.right=t.left+e.clientWidth,t.width=e.clientWidth,t.height=e.clientHeight,t.x=t.left,t.y=t.top,t}(t):re(function(e){var t,n=p(e),r=i(e),o=null==(t=e.ownerDocument)?void 0:t.body,a=F(n.scrollWidth,n.clientWidth,o?o.scrollWidth:0,o?o.clientWidth:0),c=F(n.scrollHeight,n.clientHeight,o?o.scrollHeight:0,o?o.clientHeight:0),s=-r.scrollLeft+f(e),u=-r.scrollTop;return"rtl"===d(o||n).direction&&(s+=F(n.clientWidth,o?o.clientWidth:0)-a),{width:a,height:c,x:s,y:u}}(p(e)))}function ie(e,t,n){var r="clippingParents"===t?function(e){var t=b(h(e)),n=["absolute","fixed"].indexOf(d(e).position)>=0&&c(e)?x(e):e;return a(n)?t.filter((function(e){return a(e)&&ne(e,n)&&"body"!==u(e)})):[]}(e):[].concat(t),o=[].concat(r,[n]),i=o[0],s=o.reduce((function(t,n){var r=oe(e,n);return t.top=F(r.top,t.top),t.right=K(r.right,t.right),t.bottom=K(r.bottom,t.bottom),t.left=F(r.left,t.left),t}),oe(e,i));return s.width=s.right-s.left,s.height=s.bottom-s.top,s.x=s.left,s.y=s.top,s}function ae(e){return Object.assign({},{top:0,right:0,bottom:0,left:0},e)}function ce(e,t){return t.reduce((function(t,n){return t[n]=e,t}),{})}function se(e,t){void 0===t&&(t={});var n=t,o=n.placement,i=void 0===o?e.placement:o,c=n.boundary,s=void 0===c?"clippingParents":c,u=n.rootBoundary,f=void 0===u?D:u,d=n.elementContext,l=void 0===d?_:d,v=n.altBoundary,m=void 0!==v&&v,h=n.padding,g=void 0===h?0:h,b=ae("number"!=typeof g?g:ce(g,C)),y=l===_?"reference":_,w=e.elements.reference,x=e.rects.popper,E=e.elements[m?y:l],A=ie(a(E)?E:E.contextElement||p(e.elements.popper),s,f),T=r(w),S=W({reference:T,element:x,strategy:"absolute",placement:i}),L=re(Object.assign({},x,S)),M=l===_?L:T,I={top:A.top-M.top+b.top,bottom:M.bottom-A.bottom+b.bottom,left:A.left-M.left+b.left,right:M.right-A.right+b.right},Q=e.modifiersData.offset;if(l===_&&Q){var z=Q[i];Object.keys(I).forEach((function(e){var t=[j,O].indexOf(e)>=0?1:-1,n=[k,O].indexOf(e)>=0?"y":"x";I[e]+=z[n]*t}))}return I}function ue(e,t,n){return F(e,K(t,n))}function pe(e,t,n){return void 0===n&&(n={x:0,y:0}),{top:e.top-t.height-n.y,right:e.right-t.width+n.x,bottom:e.bottom-t.height+n.y,left:e.left-t.width-n.x}}function fe(e){return[k,j,O,E].some((function(t){return e[t]>=0}))}var de=B({defaultModifiers:[V,q,{name:"computeStyles",enabled:!0,phase:"beforeWrite",fn:function(e){var t=e.state,n=e.options,r=n.gpuAcceleration,o=void 0===r||r,i=n.adaptive,a=void 0===i||i,c=n.roundOffsets,s=void 0===c||c,u={placement:N(t.placement),popper:t.elements.popper,popperRect:t.rects.popper,gpuAcceleration:o};null!=t.modifiersData.popperOffsets&&(t.styles.popper=Object.assign({},t.styles.popper,G(Object.assign({},u,{offsets:t.modifiersData.popperOffsets,position:t.options.strategy,adaptive:a,roundOffsets:s})))),null!=t.modifiersData.arrow&&(t.styles.arrow=Object.assign({},t.styles.arrow,G(Object.assign({},u,{offsets:t.modifiersData.arrow,position:"absolute",adaptive:!1,roundOffsets:s})))),t.attributes.popper=Object.assign({},t.attributes.popper,{"data-popper-placement":t.placement})},data:{}},Z,{name:"offset",enabled:!0,phase:"main",requires:["popperOffsets"],fn:function(e){var t=e.state,n=e.options,r=e.name,o=n.offset,i=void 0===o?[0,0]:o,a=M.reduce((function(e,n){return e[n]=function(e,t,n){var r=N(e),o=[E,k].indexOf(r)>=0?-1:1,i="function"==typeof n?n(Object.assign({},t,{placement:e})):n,a=i[0],c=i[1];return a=a||0,c=(c||0)*o,[E,j].indexOf(r)>=0?{x:c,y:a}:{x:a,y:c}}(n,t.rects,i),e}),{}),c=a[t.placement],s=c.x,u=c.y;null!=t.modifiersData.popperOffsets&&(t.modifiersData.popperOffsets.x+=s,t.modifiersData.popperOffsets.y+=u),t.modifiersData[r]=a}},{name:"flip",enabled:!0,phase:"main",fn:function(e){var t=e.state,n=e.options,r=e.name;if(!t.modifiersData[r]._skip){for(var o=n.mainAxis,i=void 0===o||o,a=n.altAxis,c=void 0===a||a,s=n.fallbackPlacements,u=n.padding,p=n.boundary,f=n.rootBoundary,d=n.altBoundary,l=n.flipVariations,v=void 0===l||l,m=n.allowedAutoPlacements,h=t.options.placement,g=N(h),b=s||(g===h||!v?[Y(h)]:function(e){if(N(e)===A)return[];var t=Y(e);return[te(e),t,te(t)]}(h)),y=[h].concat(b).reduce((function(e,n){return e.concat(N(n)===A?function(e,t){void 0===t&&(t={});var n=t,r=n.placement,o=n.boundary,i=n.rootBoundary,a=n.padding,c=n.flipVariations,s=n.allowedAutoPlacements,u=void 0===s?M:s,p=H(r),f=p?c?L:L.filter((function(e){return H(e)===p})):C,d=f.filter((function(e){return u.indexOf(e)>=0}));0===d.length&&(d=f);var l=d.reduce((function(t,n){return t[n]=se(e,{placement:n,boundary:o,rootBoundary:i,padding:a})[N(n)],t}),{});return Object.keys(l).sort((function(e,t){return l[e]-l[t]}))}(t,{placement:n,boundary:p,rootBoundary:f,padding:u,flipVariations:v,allowedAutoPlacements:m}):n)}),[]),w=t.rects.reference,x=t.rects.popper,S=new Map,D=!0,_=y[0],I=0;I<y.length;I++){var Q=y[I],z=N(Q),P=H(Q)===T,B=[k,O].indexOf(z)>=0,R=B?"width":"height",V=se(t,{placement:Q,boundary:p,rootBoundary:f,altBoundary:d,padding:u}),U=B?P?j:E:P?O:k;w[R]>x[R]&&(U=Y(U));var W=Y(U),q=[];if(i&&q.push(V[z]<=0),c&&q.push(V[U]<=0,V[W]<=0),q.every((function(e){return e}))){_=Q,D=!1;break}S.set(Q,q)}if(D)for(var F=function(e){var t=y.find((function(t){var n=S.get(t);if(n)return n.slice(0,e).every((function(e){return e}))}));if(t)return _=t,"break"},K=v?3:1;K>0;K--){if("break"===F(K))break}t.placement!==_&&(t.modifiersData[r]._skip=!0,t.placement=_,t.reset=!0)}},requiresIfExists:["offset"],data:{_skip:!1}},{name:"preventOverflow",enabled:!0,phase:"main",fn:function(e){var t=e.state,n=e.options,r=e.name,o=n.mainAxis,i=void 0===o||o,a=n.altAxis,c=void 0!==a&&a,s=n.boundary,u=n.rootBoundary,p=n.altBoundary,f=n.padding,d=n.tether,l=void 0===d||d,v=n.tetherOffset,h=void 0===v?0:v,g=se(t,{boundary:s,rootBoundary:u,padding:f,altBoundary:p}),b=N(t.placement),y=H(t.placement),w=!y,A=U(b),C="x"===A?"y":"x",S=t.modifiersData.popperOffsets,D=t.rects.reference,_=t.rects.popper,L="function"==typeof h?h(Object.assign({},t.rects,{placement:t.placement})):h,M={x:0,y:0};if(S){if(i||c){var I="y"===A?k:E,Q="y"===A?O:j,z="y"===A?"height":"width",P=S[A],B=S[A]+g[I],R=S[A]-g[Q],V=l?-_[z]/2:0,W=y===T?D[z]:_[z],q=y===T?-_[z]:-D[z],$=t.elements.arrow,J=l&&$?m($):{width:0,height:0},G=t.modifiersData["arrow#persistent"]?t.modifiersData["arrow#persistent"].padding:{top:0,right:0,bottom:0,left:0},Z=G[I],X=G[Q],Y=ue(0,D[z],J[z]),ee=w?D[z]/2-V-Y-Z-L:W-Y-Z-L,te=w?-D[z]/2+V+Y+X+L:q+Y+X+L,ne=t.elements.arrow&&x(t.elements.arrow),re=ne?"y"===A?ne.clientTop||0:ne.clientLeft||0:0,oe=t.modifiersData.offset?t.modifiersData.offset[t.placement][A]:0,ie=S[A]+ee-oe-re,ae=S[A]+te-oe;if(i){var ce=ue(l?K(B,ie):B,P,l?F(R,ae):R);S[A]=ce,M[A]=ce-P}if(c){var pe="x"===A?k:E,fe="x"===A?O:j,de=S[C],le=de+g[pe],ve=de-g[fe],me=ue(l?K(le,ie):le,de,l?F(ve,ae):ve);S[C]=me,M[C]=me-de}}t.modifiersData[r]=M}},requiresIfExists:["offset"]},{name:"arrow",enabled:!0,phase:"main",fn:function(e){var t,n=e.state,r=e.name,o=e.options,i=n.elements.arrow,a=n.modifiersData.popperOffsets,c=N(n.placement),s=U(c),u=[E,j].indexOf(c)>=0?"height":"width";if(i&&a){var p=function(e,t){return ae("number"!=typeof(e="function"==typeof e?e(Object.assign({},t.rects,{placement:t.placement})):e)?e:ce(e,C))}(o.padding,n),f=m(i),d="y"===s?k:E,l="y"===s?O:j,v=n.rects.reference[u]+n.rects.reference[s]-a[s]-n.rects.popper[u],h=a[s]-n.rects.reference[s],g=x(i),b=g?"y"===s?g.clientHeight||0:g.clientWidth||0:0,y=v/2-h/2,w=p[d],A=b-f[u]-p[l],T=b/2-f[u]/2+y,S=ue(w,T,A),D=s;n.modifiersData[r]=((t={})[D]=S,t.centerOffset=S-T,t)}},effect:function(e){var t=e.state,n=e.options.element,r=void 0===n?"[data-popper-arrow]":n;null!=r&&("string"!=typeof r||(r=t.elements.popper.querySelector(r)))&&ne(t.elements.popper,r)&&(t.elements.arrow=r)},requires:["popperOffsets"],requiresIfExists:["preventOverflow"]},{name:"hide",enabled:!0,phase:"main",requiresIfExists:["preventOverflow"],fn:function(e){var t=e.state,n=e.name,r=t.rects.reference,o=t.rects.popper,i=t.modifiersData.preventOverflow,a=se(t,{elementContext:"reference"}),c=se(t,{altBoundary:!0}),s=pe(a,r),u=pe(c,o,i),p=fe(s),f=fe(u);t.modifiersData[n]={referenceClippingOffsets:s,popperEscapeOffsets:u,isReferenceHidden:p,hasPopperEscaped:f},t.attributes.popper=Object.assign({},t.attributes.popper,{"data-popper-reference-hidden":p,"data-popper-escaped":f})}}]}),le="tippy-content",ve="tippy-backdrop",me="tippy-arrow",he="tippy-svg-arrow",ge={passive:!0,capture:!0};function be(e,t,n){if(Array.isArray(e)){var r=e[t];return null==r?Array.isArray(n)?n[t]:n:r}return e}function ye(e,t){var n={}.toString.call(e);return 0===n.indexOf("[object")&&n.indexOf(t+"]")>-1}function we(e,t){return"function"==typeof e?e.apply(void 0,t):e}function xe(e,t){return 0===t?e:function(r){clearTimeout(n),n=setTimeout((function(){e(r)}),t)};var n}function ke(e){return[].concat(e)}function Oe(e,t){-1===e.indexOf(t)&&e.push(t)}function je(e){return e.split("-")[0]}function Ee(e){return[].slice.call(e)}function Ae(){return document.createElement("div")}function Ce(e){return["Element","Fragment"].some((function(t){return ye(e,t)}))}function Te(e){return ye(e,"MouseEvent")}function Se(e){return!(!e||!e._tippy||e._tippy.reference!==e)}function De(e){return Ce(e)?[e]:function(e){return ye(e,"NodeList")}(e)?Ee(e):Array.isArray(e)?e:Ee(document.querySelectorAll(e))}function _e(e,t){e.forEach((function(e){e&&(e.style.transitionDuration=t+"ms")}))}function Le(e,t){e.forEach((function(e){e&&e.setAttribute("data-state",t)}))}function Me(e){var t,n=ke(e)[0];return(null==n||null==(t=n.ownerDocument)?void 0:t.body)?n.ownerDocument:document}function Ie(e,t,n){var r=t+"EventListener";["transitionend","webkitTransitionEnd"].forEach((function(t){e[r](t,n)}))}var Qe={isTouch:!1},ze=0;function Pe(){Qe.isTouch||(Qe.isTouch=!0,window.performance&&document.addEventListener("mousemove",Be))}function Be(){var e=performance.now();e-ze<20&&(Qe.isTouch=!1,document.removeEventListener("mousemove",Be)),ze=e}function Re(){var e=document.activeElement;if(Se(e)){var t=e._tippy;e.blur&&!t.state.isVisible&&e.blur()}}var Ve="undefined"!=typeof window&&"undefined"!=typeof document?navigator.userAgent:"",Ne=/MSIE |Trident\//.test(Ve);var He={animateFill:!1,followCursor:!1,inlinePositioning:!1,sticky:!1},Ue=Object.assign({appendTo:function(){return document.body},aria:{content:"auto",expanded:"auto"},delay:0,duration:[300,250],getReferenceClientRect:null,hideOnClick:!0,ignoreAttributes:!1,interactive:!1,interactiveBorder:2,interactiveDebounce:0,moveTransition:"",offset:[0,10],onAfterUpdate:function(){},onBeforeUpdate:function(){},onCreate:function(){},onDestroy:function(){},onHidden:function(){},onHide:function(){},onMount:function(){},onShow:function(){},onShown:function(){},onTrigger:function(){},onUntrigger:function(){},onClickOutside:function(){},placement:"top",plugins:[],popperOptions:{},render:null,showOnCreate:!1,touch:!0,trigger:"mouseenter focus",triggerTarget:null},He,{},{allowHTML:!1,animation:"fade",arrow:!0,content:"",inertia:!1,maxWidth:350,role:"tooltip",theme:"",zIndex:9999}),We=Object.keys(Ue);function qe(e){var t=(e.plugins||[]).reduce((function(t,n){var r=n.name,o=n.defaultValue;return r&&(t[r]=void 0!==e[r]?e[r]:o),t}),{});return Object.assign({},e,{},t)}function Fe(e,t){var n=Object.assign({},t,{content:we(t.content,[e])},t.ignoreAttributes?{}:function(e,t){return(t?Object.keys(qe(Object.assign({},Ue,{plugins:t}))):We).reduce((function(t,n){var r=(e.getAttribute("data-tippy-"+n)||"").trim();if(!r)return t;if("content"===n)t[n]=r;else try{t[n]=JSON.parse(r)}catch(e){t[n]=r}return t}),{})}(e,t.plugins));return n.aria=Object.assign({},Ue.aria,{},n.aria),n.aria={expanded:"auto"===n.aria.expanded?t.interactive:n.aria.expanded,content:"auto"===n.aria.content?t.interactive?null:"describedby":n.aria.content},n}function Ke(e,t){e.innerHTML=t}function $e(e){var t=Ae();return!0===e?t.className=me:(t.className=he,Ce(e)?t.appendChild(e):Ke(t,e)),t}function Je(e,t){Ce(t.content)?(Ke(e,""),e.appendChild(t.content)):"function"!=typeof t.content&&(t.allowHTML?Ke(e,t.content):e.textContent=t.content)}function Ge(e){var t=e.firstElementChild,n=Ee(t.children);return{box:t,content:n.find((function(e){return e.classList.contains(le)})),arrow:n.find((function(e){return e.classList.contains(me)||e.classList.contains(he)})),backdrop:n.find((function(e){return e.classList.contains(ve)}))}}function Ze(e){var t=Ae(),n=Ae();n.className="tippy-box",n.setAttribute("data-state","hidden"),n.setAttribute("tabindex","-1");var r=Ae();function o(n,r){var o=Ge(t),i=o.box,a=o.content,c=o.arrow;r.theme?i.setAttribute("data-theme",r.theme):i.removeAttribute("data-theme"),"string"==typeof r.animation?i.setAttribute("data-animation",r.animation):i.removeAttribute("data-animation"),r.inertia?i.setAttribute("data-inertia",""):i.removeAttribute("data-inertia"),i.style.maxWidth="number"==typeof r.maxWidth?r.maxWidth+"px":r.maxWidth,r.role?i.setAttribute("role",r.role):i.removeAttribute("role"),n.content===r.content&&n.allowHTML===r.allowHTML||Je(a,e.props),r.arrow?c?n.arrow!==r.arrow&&(i.removeChild(c),i.appendChild($e(r.arrow))):i.appendChild($e(r.arrow)):c&&i.removeChild(c)}return r.className=le,r.setAttribute("data-state","hidden"),Je(r,e.props),t.appendChild(n),n.appendChild(r),o(e.props,e.props),{popper:t,onUpdate:o}}Ze.$$tippy=!0;var Xe=1,Ye=[],et=[];function tt(e,t){var n,r,o,i,a,c,s,u,p,f=Fe(e,Object.assign({},Ue,{},qe((n=t,Object.keys(n).reduce((function(e,t){return void 0!==n[t]&&(e[t]=n[t]),e}),{}))))),d=!1,l=!1,v=!1,m=!1,h=[],g=xe($,f.interactiveDebounce),b=Xe++,y=(p=f.plugins).filter((function(e,t){return p.indexOf(e)===t})),w={id:b,reference:e,popper:Ae(),popperInstance:null,props:f,state:{isEnabled:!0,isVisible:!1,isDestroyed:!1,isMounted:!1,isShown:!1},plugins:y,clearDelayTimeouts:function(){clearTimeout(r),clearTimeout(o),cancelAnimationFrame(i)},setProps:function(t){0;if(w.state.isDestroyed)return;I("onBeforeUpdate",[w,t]),F();var n=w.props,r=Fe(e,Object.assign({},w.props,{},t,{ignoreAttributes:!0}));w.props=r,q(),n.interactiveDebounce!==r.interactiveDebounce&&(P(),g=xe($,r.interactiveDebounce));n.triggerTarget&&!r.triggerTarget?ke(n.triggerTarget).forEach((function(e){e.removeAttribute("aria-expanded")})):r.triggerTarget&&e.removeAttribute("aria-expanded");z(),M(),O&&O(n,r);w.popperInstance&&(X(),ee().forEach((function(e){requestAnimationFrame(e._tippy.popperInstance.forceUpdate)})));I("onAfterUpdate",[w,t])},setContent:function(e){w.setProps({content:e})},show:function(){0;var e=w.state.isVisible,t=w.state.isDestroyed,n=!w.state.isEnabled,r=Qe.isTouch&&!w.props.touch,o=be(w.props.duration,0,Ue.duration);if(e||t||n||r)return;if(S().hasAttribute("disabled"))return;if(I("onShow",[w],!1),!1===w.props.onShow(w))return;w.state.isVisible=!0,T()&&(k.style.visibility="visible");M(),N(),w.state.isMounted||(k.style.transition="none");if(T()){var i=_(),a=i.box,c=i.content;_e([a,c],0)}s=function(){var e;if(w.state.isVisible&&!m){if(m=!0,k.offsetHeight,k.style.transition=w.props.moveTransition,T()&&w.props.animation){var t=_(),n=t.box,r=t.content;_e([n,r],o),Le([n,r],"visible")}Q(),z(),Oe(et,w),null==(e=w.popperInstance)||e.forceUpdate(),w.state.isMounted=!0,I("onMount",[w]),w.props.animation&&T()&&function(e,t){U(e,t)}(o,(function(){w.state.isShown=!0,I("onShown",[w])}))}},function(){var e,t=w.props.appendTo,n=S();e=w.props.interactive&&t===Ue.appendTo||"parent"===t?n.parentNode:we(t,[n]);e.contains(k)||e.appendChild(k);X(),!1}()},hide:function(){0;var e=!w.state.isVisible,t=w.state.isDestroyed,n=!w.state.isEnabled,r=be(w.props.duration,1,Ue.duration);if(e||t||n)return;if(I("onHide",[w],!1),!1===w.props.onHide(w))return;w.state.isVisible=!1,w.state.isShown=!1,m=!1,d=!1,T()&&(k.style.visibility="hidden");if(P(),H(),M(),T()){var o=_(),i=o.box,a=o.content;w.props.animation&&(_e([i,a],r),Le([i,a],"hidden"))}Q(),z(),w.props.animation?T()&&function(e,t){U(e,(function(){!w.state.isVisible&&k.parentNode&&k.parentNode.contains(k)&&t()}))}(r,w.unmount):w.unmount()},hideWithInteractivity:function(e){0;D().addEventListener("mousemove",g),Oe(Ye,g),g(e)},enable:function(){w.state.isEnabled=!0},disable:function(){w.hide(),w.state.isEnabled=!1},unmount:function(){0;w.state.isVisible&&w.hide();if(!w.state.isMounted)return;Y(),ee().forEach((function(e){e._tippy.unmount()})),k.parentNode&&k.parentNode.removeChild(k);et=et.filter((function(e){return e!==w})),w.state.isMounted=!1,I("onHidden",[w])},destroy:function(){0;if(w.state.isDestroyed)return;w.clearDelayTimeouts(),w.unmount(),F(),delete e._tippy,w.state.isDestroyed=!0,I("onDestroy",[w])}};if(!f.render)return w;var x=f.render(w),k=x.popper,O=x.onUpdate;k.setAttribute("data-tippy-root",""),k.id="tippy-"+w.id,w.popper=k,e._tippy=w,k._tippy=w;var j=y.map((function(e){return e.fn(w)})),E=e.hasAttribute("aria-expanded");return q(),z(),M(),I("onCreate",[w]),f.showOnCreate&&te(),k.addEventListener("mouseenter",(function(){w.props.interactive&&w.state.isVisible&&w.clearDelayTimeouts()})),k.addEventListener("mouseleave",(function(e){w.props.interactive&&w.props.trigger.indexOf("mouseenter")>=0&&(D().addEventListener("mousemove",g),g(e))})),w;function A(){var e=w.props.touch;return Array.isArray(e)?e:[e,0]}function C(){return"hold"===A()[0]}function T(){var e;return!!(null==(e=w.props.render)?void 0:e.$$tippy)}function S(){return u||e}function D(){var e=S().parentNode;return e?Me(e):document}function _(){return Ge(k)}function L(e){return w.state.isMounted&&!w.state.isVisible||Qe.isTouch||a&&"focus"===a.type?0:be(w.props.delay,e?0:1,Ue.delay)}function M(){k.style.pointerEvents=w.props.interactive&&w.state.isVisible?"":"none",k.style.zIndex=""+w.props.zIndex}function I(e,t,n){var r;(void 0===n&&(n=!0),j.forEach((function(n){n[e]&&n[e].apply(void 0,t)})),n)&&(r=w.props)[e].apply(r,t)}function Q(){var t=w.props.aria;if(t.content){var n="aria-"+t.content,r=k.id;ke(w.props.triggerTarget||e).forEach((function(e){var t=e.getAttribute(n);if(w.state.isVisible)e.setAttribute(n,t?t+" "+r:r);else{var o=t&&t.replace(r,"").trim();o?e.setAttribute(n,o):e.removeAttribute(n)}}))}}function z(){!E&&w.props.aria.expanded&&ke(w.props.triggerTarget||e).forEach((function(e){w.props.interactive?e.setAttribute("aria-expanded",w.state.isVisible&&e===S()?"true":"false"):e.removeAttribute("aria-expanded")}))}function P(){D().removeEventListener("mousemove",g),Ye=Ye.filter((function(e){return e!==g}))}function B(e){if(!(Qe.isTouch&&(v||"mousedown"===e.type)||w.props.interactive&&k.contains(e.target))){if(S().contains(e.target)){if(Qe.isTouch)return;if(w.state.isVisible&&w.props.trigger.indexOf("click")>=0)return}else I("onClickOutside",[w,e]);!0===w.props.hideOnClick&&(w.clearDelayTimeouts(),w.hide(),l=!0,setTimeout((function(){l=!1})),w.state.isMounted||H())}}function R(){v=!0}function V(){v=!1}function N(){var e=D();e.addEventListener("mousedown",B,!0),e.addEventListener("touchend",B,ge),e.addEventListener("touchstart",V,ge),e.addEventListener("touchmove",R,ge)}function H(){var e=D();e.removeEventListener("mousedown",B,!0),e.removeEventListener("touchend",B,ge),e.removeEventListener("touchstart",V,ge),e.removeEventListener("touchmove",R,ge)}function U(e,t){var n=_().box;function r(e){e.target===n&&(Ie(n,"remove",r),t())}if(0===e)return t();Ie(n,"remove",c),Ie(n,"add",r),c=r}function W(t,n,r){void 0===r&&(r=!1),ke(w.props.triggerTarget||e).forEach((function(e){e.addEventListener(t,n,r),h.push({node:e,eventType:t,handler:n,options:r})}))}function q(){var e;C()&&(W("touchstart",K,{passive:!0}),W("touchend",J,{passive:!0})),(e=w.props.trigger,e.split(/\s+/).filter(Boolean)).forEach((function(e){if("manual"!==e)switch(W(e,K),e){case"mouseenter":W("mouseleave",J);break;case"focus":W(Ne?"focusout":"blur",G);break;case"focusin":W("focusout",G)}}))}function F(){h.forEach((function(e){var t=e.node,n=e.eventType,r=e.handler,o=e.options;t.removeEventListener(n,r,o)})),h=[]}function K(e){var t,n=!1;if(w.state.isEnabled&&!Z(e)&&!l){var r="focus"===(null==(t=a)?void 0:t.type);a=e,u=e.currentTarget,z(),!w.state.isVisible&&Te(e)&&Ye.forEach((function(t){return t(e)})),"click"===e.type&&(w.props.trigger.indexOf("mouseenter")<0||d)&&!1!==w.props.hideOnClick&&w.state.isVisible?n=!0:te(e),"click"===e.type&&(d=!n),n&&!r&&ne(e)}}function $(e){var t=e.target,n=S().contains(t)||k.contains(t);"mousemove"===e.type&&n||function(e,t){var n=t.clientX,r=t.clientY;return e.every((function(e){var t=e.popperRect,o=e.popperState,i=e.props.interactiveBorder,a=je(o.placement),c=o.modifiersData.offset;if(!c)return!0;var s="bottom"===a?c.top.y:0,u="top"===a?c.bottom.y:0,p="right"===a?c.left.x:0,f="left"===a?c.right.x:0,d=t.top-r+s>i,l=r-t.bottom-u>i,v=t.left-n+p>i,m=n-t.right-f>i;return d||l||v||m}))}(ee().concat(k).map((function(e){var t,n=null==(t=e._tippy.popperInstance)?void 0:t.state;return n?{popperRect:e.getBoundingClientRect(),popperState:n,props:f}:null})).filter(Boolean),e)&&(P(),ne(e))}function J(e){Z(e)||w.props.trigger.indexOf("click")>=0&&d||(w.props.interactive?w.hideWithInteractivity(e):ne(e))}function G(e){w.props.trigger.indexOf("focusin")<0&&e.target!==S()||w.props.interactive&&e.relatedTarget&&k.contains(e.relatedTarget)||ne(e)}function Z(e){return!!Qe.isTouch&&C()!==e.type.indexOf("touch")>=0}function X(){Y();var t=w.props,n=t.popperOptions,r=t.placement,o=t.offset,i=t.getReferenceClientRect,a=t.moveTransition,c=T()?Ge(k).arrow:null,u=i?{getBoundingClientRect:i,contextElement:i.contextElement||S()}:e,p=[{name:"offset",options:{offset:o}},{name:"preventOverflow",options:{padding:{top:2,bottom:2,left:5,right:5}}},{name:"flip",options:{padding:5}},{name:"computeStyles",options:{adaptive:!a}},{name:"$$tippy",enabled:!0,phase:"beforeWrite",requires:["computeStyles"],fn:function(e){var t=e.state;if(T()){var n=_().box;["placement","reference-hidden","escaped"].forEach((function(e){"placement"===e?n.setAttribute("data-placement",t.placement):t.attributes.popper["data-popper-"+e]?n.setAttribute("data-"+e,""):n.removeAttribute("data-"+e)})),t.attributes.popper={}}}}];T()&&c&&p.push({name:"arrow",options:{element:c,padding:3}}),p.push.apply(p,(null==n?void 0:n.modifiers)||[]),w.popperInstance=de(u,k,Object.assign({},n,{placement:r,onFirstUpdate:s,modifiers:p}))}function Y(){w.popperInstance&&(w.popperInstance.destroy(),w.popperInstance=null)}function ee(){return Ee(k.querySelectorAll("[data-tippy-root]"))}function te(e){w.clearDelayTimeouts(),e&&I("onTrigger",[w,e]),N();var t=L(!0),n=A(),o=n[0],i=n[1];Qe.isTouch&&"hold"===o&&i&&(t=i),t?r=setTimeout((function(){w.show()}),t):w.show()}function ne(e){if(w.clearDelayTimeouts(),I("onUntrigger",[w,e]),w.state.isVisible){if(!(w.props.trigger.indexOf("mouseenter")>=0&&w.props.trigger.indexOf("click")>=0&&["mouseleave","mousemove"].indexOf(e.type)>=0&&d)){var t=L(!1);t?o=setTimeout((function(){w.state.isVisible&&w.hide()}),t):i=requestAnimationFrame((function(){w.hide()}))}}else H()}}function nt(e,t){void 0===t&&(t={});var n=Ue.plugins.concat(t.plugins||[]);document.addEventListener("touchstart",Pe,ge),window.addEventListener("blur",Re);var r=Object.assign({},t,{plugins:n}),o=De(e).reduce((function(e,t){var n=t&&tt(t,r);return n&&e.push(n),e}),[]);return Ce(e)?o[0]:o}nt.defaultProps=Ue,nt.setDefaultProps=function(e){Object.keys(e).forEach((function(t){Ue[t]=e[t]}))},nt.currentInput=Qe;Object.assign({},Z,{effect:function(e){var t=e.state,n={popper:{position:t.options.strategy,left:"0",top:"0",margin:"0"},arrow:{position:"absolute"},reference:{}};Object.assign(t.elements.popper.style,n.popper),t.styles=n,t.elements.arrow&&Object.assign(t.elements.arrow.style,n.arrow)}});nt.setDefaultProps({render:Ze});const rt=nt}},t={};function n(r){var o=t[r];if(void 0!==o)return o.exports;var i=t[r]={id:r,exports:{}};return e[r](i,i.exports,n),i.exports}n.n=e=>{var t=e&&e.__esModule?()=>e.default:()=>e;return n.d(t,{a:t}),t},n.d=(e,t)=>{for(var r in t)n.o(t,r)&&!n.o(e,r)&&Object.defineProperty(e,r,{enumerable:!0,get:t[r]})},n.o=(e,t)=>Object.prototype.hasOwnProperty.call(e,t),n(108),n(278),n(502);n(861)})();
