+++ assets/mix-manifest.json
@@ -1,7 +1,7 @@
 {
     "/js/ccfw-frontend.js": "/js/ccfw-frontend.js?id=9067385904df5f204e04",
     "/js/ccfw-cookie-manage.js": "/js/ccfw-cookie-manage.js?id=00be7b92148d8571da32",
-    "/js/ccfw-admin-app.js": "/js/ccfw-admin-app.js?id=5baeddfc097efa898bdf",
+    "/js/ccfw-admin-app.js": "/js/ccfw-admin-app.js?id=0e7f85bb37e89e778665",
     "/js/ccfw-ie11CustomProperties.js": "/js/ccfw-ie11CustomProperties.js?id=76c70d392e2e369a9b33",
     "/js/ccfw-frontend-legacy.js": "/js/ccfw-frontend-legacy.js?id=acb4c327aade27609891",
     "/css/ccfw-frontend-legacy.css": "/css/ccfw-frontend-legacy.css?id=83c305900fe21064f9f3",
