+++ components/CookieManagement/src/js/modules/listeners.js
@@ -1,5 +1,6 @@
 import { CCFW } from './global';
 import { Builder } from './build';
+import { Helper } from './helper';
 
 const groupSave = (section, callback) => {
     let element = jQuery('.ccfw-' + section + '-group-save');
@@ -50,12 +51,12 @@ const debugToggle = () => {
 };
 const debugImportObject = () => {
     let element = jQuery('.' + CCFW.debug.import.textarea);
-    element.on('keypress', CCFW.manage.readonly);
+    element.on('keypress', Helper.readonly);
     element.on('keyup', importObjectDebug);
 };
 const debugCopyObject = () => {
     let element = jQuery('#' + CCFW.debug.preContainer).find('pre');
-    element.on('click', copyObjectDebug);
+    element.on('click', Helper.copyText);
 };
 
 function toggleDebug () {
@@ -76,34 +77,19 @@ function toggleDebug () {
 
 const importObjectDebug = () => {
     let content = jQuery('.' + CCFW.debug.import.textarea).val();
+    let json = Helper.isJson(content); // returns the object
 
-    if (CCFW.manage.isJson(content)) {
-        let json = JSON.parse(content);
-
+    if (json) {
         /////////
-        // could extend to check if the object matches expected sections
+        // could extend the check to detect if the object matches expected sections
         /////////
 
-        if (json) {
-            CCFW.sections = json;
-            Builder.load();
+        CCFW.sections = json;
+        Builder.load();
 
-            CCFW.debug.output();
-        }
+        CCFW.debug.output();
     }
 };
-
-function copyObjectDebug () {
-    let pre = jQuery(this);
-    let content = pre.text();
-    let input = jQuery('<input>');
-    jQuery('body').append(input);
-    input.val(content).select();
-    document.execCommand('copy');
-    input.remove();
-    pre.addClass('copied');
-    setTimeout(() => pre.removeClass('copied'), 500);
-}
 
 ///////////////////////////////////////
 ///////////////////////////////////////
